% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquamet-package.R
\docType{package}
\name{aquamet-package}
\alias{aquamet-package}
\title{aquamet: Calculate NRSA and NLA physical habitat metrics}
\description{
This package calculates physical habitat metrics used in the National Rivers and Streams Assessment and the National Lakes Assessment. It also contains functions to assign condition class for physical habitat for NRSA and NLA.
}
\keyword{internal}
