% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquametAlt.r
\docType{package}
\name{aquametAlt}
\alias{aquametAlt}
\alias{aquametAlt-package}
\title{aquametAlt: Calculate metrics used in NRSA and NLA}
\description{
aquametAlt: Calculate metrics used in NRSA and NLA
}

