% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignFishCondition.r
\name{assignFishCondition}
\alias{assignFishCondition}
\title{Assign fish assemblage condition}
\usage{
assignFishCondition(inMMI, sampID = "UID", ecoreg = "ECOREG",
  mmi = "MMI_FISH", wsarea = "WSAREA", totlnind = "TOTLNIND")
}
\arguments{
\item{inMMI}{Input data frame which must contain the 
variables named in the remaining arguments.}

\item{sampID}{A character vector containing the names of all 
variables in inMets that specify a unique sample. If not specified, 
the default is \emph{UID}}

\item{ecoreg}{A string with the name of the ecoregion variable. 
Valid values that correspond to regions used in NRSA are
CPL, NAP, NPL, SAP, sPL, TPL, UMW, WMT, and XER.}

\item{mmi}{A string with the name of the MMI input variable. The
default is 'MMI_FISH' because this is the name of the output 
from the \code{calcFishMMI()} function.}

\item{wsarea}{A string with the name of the watershed area variable.
Watershed area is assumed to be in square km. This variable is required 
in sites where fish were sampled but none were collected, along with
\emph{totlnind}. For these sites, \emph{mmi} should be missing and
\emph{totlnind} should be set to 0. At sites with a \emph{wsarea} <=2 
km2, fish are not expected and thus, condition is set as Not Assessed.}

\item{totlnind}{A string with the name of the variable in 
\emph{inMMI} with the number of individuals in the fish sample. 
This variable is necessary to assign condition in sites where 
fish were sampled but none were collected. 
In such cases \emph{totlnind} should be 0 and \emph{mmi} missing.}
}
\value{
A data frame containing all of the input variables and the 
variable FISH_MMI_COND for each site.
}
\description{
Using ecoregion, fish MMI score, 
log10(watershed area), and total number of individuals 
to assign condition using NRSA thresholds.
}
\details{
If site is not sampled for fish, \emph{mmi} and \emph{totlnind} 
should be missing values in the \emph{inMMI} data frame.
}
\examples{
{
 fishMMIex <- data.frame(SITE_ID = c('Site1','Site2','Site3','Site4',
                                   'Site5','Site1','Site6','Site7'),
 VISIT_NO = c(1,1,1,1,1,2,1,1), ECO9 = c('CPL','XER','NAP','WMT',
                                         'TPL','CPL','SAP','SAP'),
 MMI_FISH = c(57.3, 79.2, 63.2, 73.2, NA, 53.1, NA, NA), 
 WSAREAKM = c(10, 56, 67, 54, 1.3, 10, 3.4, 2.6), 
 TOTLNIND = c(346, 230, 507, 304, 0, 209, 0, NA), stringsAsFactors=FALSE) 
 
 fishCond <- assignFishCondition(fishMMIex, sampID = c('SITE_ID','VISIT_NO'),
 ecoreg = 'ECO9', mmi = 'MMI_FISH', wsarea = 'WSAREAKM',
 totlnind = 'TOTLNIND')
 
 fishCond
}

}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
\keyword{survey}
