% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAllFishMets.r
\name{calcAllFishMets}
\alias{calcAllFishMets}
\title{Calculate all fish metrics}
\usage{
calcAllFishMets(indf, inTaxa = NULL, sampID = "UID", dist = "IS_DISTINCT",
  ct = "TOTAL", anomct = "ANOM_CT", taxa_id = "TAXA_ID",
  tol = "TOLERANCE", tolval = "TOL_VAL", vel = "VELOCITY",
  habitat = "HABITAT", trophic = "TROPHIC", migr = "MIGRATORY",
  nonnat = "NONNATIVE", reprod = "REPROD", temp = "TEMP",
  family = "FAMILY", genus = "GENUS", comname = "NAME")
}
\arguments{
\item{indf}{A data frame containing, at minimum, the variables 
specified in the arguments for sampID, dist, ct, and taxa_id}

\item{inTaxa}{a data frame containing taxonomic information, 
including variables for PHYLUM, CLASS, ORDER, FAMILY, SUBFAMILY, 
and TRIBE, as well as autecology traits with names that match those 
in the arguments ffg, habit, and ptv. In addition, there
should be a variable with the name in argument taxa_id that matches 
with all of those in the indf data frame}

\item{sampID}{A character vector containing the names of all 
variables in indf that specify a unique sample. If not specified, 
the default is \emph{UID}}

\item{dist}{A string with the name of the distinctness variable, 
which is assumed to have only values of 0 or 1. If not specified, 
the default is \emph{IS_DISTINCT}.}

\item{ct}{A string with the name of the count variable. If not 
specified, the default is \emph{FINAL_CT}.}

\item{anomct}{A string with the name of the count variable. If not 
specified, the default is \emph{ANOM_CT}.}

\item{taxa_id}{A string with the name of the taxon ID variable 
in \emph{indf} that matches that in \emph{inTaxa}. The default 
value is \emph{TAXA_ID}.}

\item{tol}{A string with the name of the variable in the 
\emph{inTaxa} taxalist containing tolerance categories. Valid
values include S (sensitive), I (intermediate), and T (tolerant).
The default value is \emph{TOLERANCE}.}

\item{tolval}{A string with the name of the variable in the 
\emph{inTaxa} taxalist containing numeric tolerance values.
Valid values range from 0 to 10. The default value is TOL_VAL.}

\item{vel}{A strings with the name of the variable in the 
\emph{inTaxa} taxalist containing velocity preference values.
Valid values include R (Rheophil), P (Pool), O (Other), or 
blank, if unknown. The default value is \emph{VELOCITY}.}

\item{habitat}{A string with the name of the variable in 
\emph{inTaxa} containing habitat preference values. Valid
values include B (benthic), W (water column), E (edge), or
blank, if unknown. The default value is \emph{HABITAT}.}

\item{trophic}{A string with the name of the variable in
\emph{inTaxa} containing trophic category values. Valid
values include I (invertivore), C (carnivore), O (omnivore),
H (herbivore), or blank if unknown. The default value is
\emph{inTaxa} is TROPHIC.}

\item{migr}{A string with the name of the variable in 
\emph{inTaxa} containing migratory status value. Valid 
values include N (No), Y (Yes), and blank if unknown. The
default value is \emph{MIGRATORY}.}

\item{nonnat}{A string with the name of the optional variable in 
\emph{inCts} containing non-native status. Valid values are 'Y' for 
non-native and 'N' for native. The default name 
is \emph{NONNATIVE}.}

\item{reprod}{A string with the name of the variable in 
\emph{inTaxa} containing reproductive trait values. Valid 
values include C (Clean, coarse lithophil), D (Drifter),
G (Guarder), O (Other), or blank if unknown. The default
value is \emph{REPROD}.}

\item{temp}{A string with the name of the variable in 
\emph{inTaxa} containing the temperature preference values.
Valid values include WM (warm), CD (cold water), CL (cool water), 
or blank if unknow. The default value is \emph{TEMP}.}

\item{family}{A string with the name of the variable in the 
\emph{inTaxa} taxalist containing family name. The default value 
is \emph{FAMILY}.}

\item{genus}{A string with the name of the variable in the 
\emph{inTaxa} taxalist containing the genus name. The default
value is \emph{GENUS}}

\item{comname}{A string with the name of the variable in the 
\emph{inTaxa} taxalist containing the common name. The 
default value is \emph{NAME}.}
}
\value{
A data frame containing the variables in sampID and 
all fish metrics as additional variables. These are listed in
the document \emph{NRSA_Fish_Metric_Descriptions.pdf} included
in this package, which contains metric descriptions.
}
\description{
This is a wrapper function that calculates 
all fish metrics as used in the National Aquatic 
Resource Surveys.
}
\examples{
  \dontrun{
  data(fishEx)
  head(fishEx)
  head(fishTaxa)
  # Calculate metrics for bentIn, using the taxonomy in the count file as is
  fishMetrics <- calcAllFishMets(indf=fishEx, inTaxa=fishTaxa,
                     sampID='UID',dist='IS_DISTINCT',
                     ct='FINAL_CT',anomct='ANOM_CT',taxa_id='TAXA_ID',
                     tol='TOLERANCE_NRSA',tolval='TOL_VAL_EMAPW',vel='VEL_NRSA',
                     habitat='HABITAT_NRSA',trophic='TROPHIC_NRSA',
                     migr='MIGR_NRSA',nonnat='NON_NATIVE',reprod='REPROD_NRSA',
                     temp='TEMP_NRSA',family='FAMILY',genus='GENUS',
                     comname='FINAL_NAME')
                     head(fishMetrics)}
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
\keyword{survey}
