% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBentMetricSubsets.r
\name{calcBentFFGmets}
\alias{calcBentFFGmets}
\title{Calculate functional feeding group benthic metrics}
\usage{
calcBentFFGmets(inCts, inTaxa, sampID = "UID", dist = "IS_DISTINCT",
  ct = "TOTAL", taxa_id = "TAXA_ID", ffg = "FFG")
}
\arguments{
\item{inCts}{A data frame containing, at minimum, the variables 
specified in the arguments for sampID, dist, ct, and taxa_id}

\item{inTaxa}{a data frame containing taxonomic information, 
including variables for PHYLUM, CLASS, ORDER, FAMILY, SUBFAMILY, 
and TRIBE, as well as autecology traits with names that match those 
in the arguments ffg, habit, and ptv. In addition, there
should be a variable with the name in argument taxa_id that matches 
with all of those in the indf data frame}

\item{sampID}{A character vector containing the names of all 
variables in indf that specify a unique sample. If not specified, 
the default is \emph{UID}}

\item{dist}{A string with the name of the distinctness variable, 
which is assumed to have only values of 0 or 1. If not specified, 
the default is \emph{IS_DISTINCT}.}

\item{ct}{A string with the name of the count variable. If not 
specified, the default is \emph{TOTAL}.}

\item{taxa_id}{A string with the name of the taxon ID variable 
in \emph{inCts} that matches that in \emph{inTaxa}. The default 
value is \emph{TAXA_ID}.}

\item{ffg}{A string with the name of the functional feeding group 
variable in inTaxa. The default value is \emph{FFG}. Values used
in calculations include CF, CG, PR, SH, Sc, representing 
collector-filterer, collector-gatherer, predator, shredder, and
scraper, respectively. Each taxon may have more than 
one ffg value.}
}
\value{
A data frame containing the variables in sampID and 
the benthic macroinvertebrate metrics as additional variables. 
Metrics included are named COFINTAX, COFIPIND, COFIPTAX,
COFITRICNTAX, COFITRICPIND, COFITRICPTAX, COGANTAX, COGAPIND,
COGAPTAX, PREDNTAX, PREDPIND, PREDPTAX, SCRPNTAX, SCRPPIND,
SCRPPTAX, SHRDNTAX, SHRDPIND, and SHRDPTAX.

Metric descriptions are included in \emph{NRSA_Invertebrate_Metric_Descriptions.pdf},
included in this package.
}
\description{
This function calculates 
all functional feeding group benthic metrics using trait 
information supplied.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
