% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBentMetricSubsets.r
\name{calcBentHabitMets}
\alias{calcBentHabitMets}
\title{Calculate habit benthic metrics}
\usage{
calcBentHabitMets(inCts, inTaxa, sampID = "UID", dist = "IS_DISTINCT",
  ct = "TOTAL", taxa_id = "TAXA_ID", habit = "HABIT")
}
\arguments{
\item{inCts}{A data frame containing, at minimum, the variables 
specified in the arguments for sampID, dist, ct, and taxa_id}

\item{inTaxa}{a data frame containing taxonomic information, 
including variables for PHYLUM, CLASS, ORDER, FAMILY, SUBFAMILY, 
and TRIBE, as well as autecology traits with names that match those 
in the arguments ffg, habit, and ptv. In addition, there
should be a variable with the name in argument taxa_id that matches 
with all of those in the indf data frame}

\item{sampID}{A character vector containing the names of all 
variables in indf that specify a unique sample. If not specified, 
the default is \emph{UID}}

\item{dist}{A string with the name of the distinctness variable, 
which is assumed to have only values of 0 or 1. If not specified, 
the default is \emph{IS_DISTINCT}.}

\item{ct}{A string with the name of the count variable. If not 
specified, the default is \emph{TOTAL}.}

\item{taxa_id}{A string with the name of the taxon ID variable 
in \emph{inCts} that matches that in \emph{inTaxa}. The default 
value is \emph{TAXA_ID}.}

\item{habit}{A string with the name of the habit variable in inTaxa. 
The default value is \emph{HABIT}. Values for habit that are used in
calculations include BU, CB, CN, SP, SW, representing burrower, 
climber, clinger, sprawler, and swimmer, respectively. Each taxon 
may have more than one value for habit.}
}
\value{
A data frame containing the variables in sampID and 
the benthic macroinvertebrate metrics as additional variables.
 The metric names include  BURRNTAX, BURRPIND, BURRPTAX, CLMBNTAX, 
 CLMBPIND, CLMBPTAX, CLNGNTAX, CLNGPIND, CLNGPTAX, SPWLNTAX, 
 SPWLPIND, SPWLPTAX, SWIMNTAX, SWIMPIND, and SWIMPTAX. 
 
Metric descriptions are included in \emph{NRSA_Invertebrate_Metric_Descriptions.pdf},
included in this package.
}
\description{
This function calculates habit 
benthic metrics.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
