% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBentMetricSubsets.r
\name{calcBentTaxMets}
\alias{calcBentTaxMets}
\title{Calculate taxonomy metrics for benthic macroinvertebrates}
\usage{
calcBentTaxMets(inCts, inTaxa, sampID = "UID", dist = "IS_DISTINCT",
  ct = "TOTAL", taxa_id = "TAXA_ID")
}
\arguments{
\item{inCts}{A data frame containing, at minimum, the variables 
specified in the arguments for sampID, dist, ct, and taxa_id}

\item{inTaxa}{a data frame containing taxonomic information, 
including variables for PHYLUM, CLASS, ORDER, FAMILY, SUBFAMILY, 
and TRIBE, as well as autecology traits with names that match those 
in the arguments ffg, habit, and ptv. In addition, there
should be a variable with the name in argument taxa_id that matches 
with all of those in the indf data frame}

\item{sampID}{A character vector containing the names of all 
variables in indf that specify a unique sample. If not specified, 
the default is \emph{UID}}

\item{dist}{A string with the name of the distinctness variable, 
which is assumed to have only values of 0 or 1. If not specified, 
the default is \emph{IS_DISTINCT}.}

\item{ct}{A string with the name of the count variable. If not 
specified, the default is \emph{TOTAL}.}

\item{taxa_id}{A string with the name of the taxon ID variable 
in \emph{inCts} that matches that in \emph{inTaxa}. The default 
value is \emph{TAXA_ID}.}
}
\value{
A data frame containing the variables in sampID and 
the benthic macroinvertebrate metrics as additional variables.
 These metrics are named  AMPHNTAX, AMPHPIND, AMPHPTAX, CHIRNTAX, 
 CHIRPIND, CHIRPTAX, CRUSNTAX, CRUSPIND, CRUSPTAX, DIPTNTAX,
 DIPTPIND, DIPTPTAX, EPHENTAX, EPHEPIND, EPHEPTAX, EPOTNTAX, 
 EPOTPIND, EPOTPTAX, EPT_NTAX, EPT_PIND, EPT_PTAX, HEMINTAX, 
 HEMIPIND, HEMIPTAX, MITENTAX, MITEPIND, MITEPTAX, MOLLNTAX, 
 MOLLPIND, MOLLPTAX, NOINNTAX, NOINPIND, NOINPTAX, ODONNTAX, 
 ODONPIND, ODONPTAX, OLLENTAX, OLLEPIND, OLLEPTAX, ORTHNTAX, 
 ORTHPIND, ORTHPTAX, PLECNTAX, PLECPIND, PLECPTAX, TANYNTAX, 
 TANYPIND, TANYPTAX, TRICNTAX, TRICPIND, TRICPTAX, TUBINAIDNTAX, 
 TUBINAIDPIND, TUBINAIDPTAX, and ORTHCHIRPIND. 

Metric descriptions are included in \emph{NRSA_Invertebrate_Metric_Descriptions.pdf},
included in this package.
}
\description{
This function calculates all taxonomy related 
benthic metrics.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
