% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBentMetricSubsets.r
\name{calcBentTolMets}
\alias{calcBentTolMets}
\title{Calculate tolerance benthic metrics}
\usage{
calcBentTolMets(inCts, inTaxa, sampID = "UID", dist = "IS_DISTINCT",
  ct = "TOTAL", taxa_id = "TAXA_ID", ptv = "PTV")
}
\arguments{
\item{inCts}{A data frame containing, at minimum, the variables 
specified in the arguments for sampID, dist, ct, and taxa_id}

\item{inTaxa}{a data frame containing taxonomic information, 
including variables for PHYLUM, CLASS, ORDER, FAMILY, SUBFAMILY, 
and TRIBE, as well as autecology traits with names that match those 
in the arguments ffg, habit, and ptv. In addition, there
should be a variable with the name in argument taxa_id that matches 
with all of those in the indf data frame}

\item{sampID}{A character vector containing the names of all 
variables in indf that specify a unique sample. If not specified, 
the default is \emph{UID}}

\item{dist}{A string with the name of the distinctness variable, 
which is assumed to have only values of 0 or 1. If not specified, 
the default is \emph{IS_DISTINCT}.}

\item{ct}{A string with the name of the count variable. If not 
specified, the default is \emph{TOTAL}.}

\item{taxa_id}{A string with the name of the taxon ID variable 
in \emph{inCts} that matches that in \emph{inTaxa}. The default 
value is \emph{TAXA_ID}.}

\item{ptv}{A string with the name of the pollution tolerance value
variable in inTaxa. The default value is \emph{PTV}. Valid values
for ptv range from 0 to 10.}
}
\value{
A data frame containing the variables in sampID and  
the benthic macroinvertebrate metrics as additional variables. 
 The names of metrics are  FACLNTAX, FACLPIND, FACLPTAX, INTLNTAX, 
 INTLPIND, INTLPTAX, NTOLNTAX, NTOLPIND, NTOLPTAX, STOLNTAX, 
 STOLPIND, STOLPTAX, TL01NTAX, TL01PIND, TL01PTAX, TL23NTAX, 
 TL23PIND, TL23PTAX, TL45NTAX, TL45PIND, TL45PTAX, TL67NTAX, 
 TL67PIND, TL67PTAX, TOLRNTAX, TOLRPIND, TOLRPTAX, and WTD_TV.
 
Metric descriptions are included in \emph{NRSA_Invertebrate_Metric_Descriptions.pdf},
included in this package.
}
\description{
This function calculates tolerance 
benthic metrics.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
