% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFishAnomMets.r
\name{calcFishAnomMets}
\alias{calcFishAnomMets}
\title{Calculate anomaly-based metrics}
\usage{
calcFishAnomMets(indata, sampID = "UID", ct = "TOTAL", anomct = "ANOM_CT")
}
\arguments{
\item{indata}{Input data frame containing variables as identified
in the arguments for \emph{sampID}, \emph{dist}, \emph{ct}, 
\emph{taxa_id}.}

\item{sampID}{sampID A character vector containing the names of all 
variables in indf that specify a unique sample. If not specified, 
the default is \emph{UID}.
#' @param ct A string with the name of the count variable. If not 
specified, the default is \emph{TOTAL}.}

\item{ct}{A string with the name of the count variable. If not 
specified, the default is \emph{TOTAL}.}

\item{anomct}{A string with the name of the count variable. If not 
specified, the default is \emph{ANOM_CT}.}
}
\value{
A data frame containing the variables in sampID and 
the fish anomaly metric ANOMPIND as an additional variable. Metric 
descriptions are included in \emph{NRSA_Fish_Metric_Descriptions.pdf},
included in this package.
}
\description{
This function calculates all of the metrics based
only on fish anomalies.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
