% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFishMMI.r
\name{calcFishMMI}
\alias{calcFishMMI}
\title{Calculate NRSA Fish MMI}
\usage{
calcFishMMI(inMets, sampID = "UID", ecoreg = "ECOREG",
  lwsarea = "LWSAREA")
}
\arguments{
\item{inMets}{A data frame containing, at minimum, the variables 
specified in the arguments for sampID, ecoreg, lwsarea and the metrics 
necessary for calculation of MMI by region. If values for more than 
the necessary metrics are included in the input data frame, the 
unnecessary  metrics will be ignored for each given site.

The necessary metrics, by aggregate ecoregion, are:

 CPL: ALIENPIND, RBCATONTAX, LOTPIND, INTLMIGRPTAX, LITHPIND, NAT_TOTLNTAX, 
     TOLRNTAX, INVPTAX
     
 NAP: ALIENNTAX, SALMNTAX, NAT_RHEOPIND, INTLMIGRPIND, LITHPTAX,
     NTOLPTAX, TOLRNTAX, INVNTAX
     
 NPL: ALIENNTAX, NAT_CYPRPIND, LOTNTAX, MIGRNTAX, LITHPIND, NTOLPTAX,
     NAT_INTLPIND, NAT_CARNNTAX 
     
 SAP: NAT_PTAX, NAT_CENTNTAX, NAT_NTOLBENTPTAX, NAT_MIGRNTAX,
     NAT_LITHPIND, NTOLPTAX, TOLRPTAX, INVPIND
     
 SPL: NAT_PIND, CYPRPTAX, RHEOPIND, NAT_MIGRPTAX, LITHNTAX,
     NAT_NTOLNTAX, TOLRNTAX, HERBPTAX
 
 TPL: ALIENNTAX, NAT_ICTAPIND, RHEONTAX, INTLMIGRNTAX, LITHPIND,
     NAT_NTOLNTAX, INTLPTAX, CARNNTAX
     
 UMW: NAT_PTAX, CYPRNTAX, INTLLOTNTAX, INTLMIGRPTAX, LITHPIND,
     NTOLNTAX, TOLRNTAX, INTLINVPTAX
     
 WMT: NAT_PIND, NAT_CATOPIND, INTLLOTPTAX, NAT_MIGRPTAX, LITHPTAX,
     NAT_TOTLNTAX, TOLRNTAX, NAT_HERBPTAX
     
 XER: NAT_PIND, CENTPTAX, RHEOPIND, MIGRPTAX, LITHNTAX, NTOLPTAX,
     TOLRNTAX, BENTINVPTAX}

\item{sampID}{A character vector containing the names of all 
variables in inMets that specify a unique sample. If not specified, 
the default is \emph{UID}}

\item{ecoreg}{A string with the name of the ecoregion variable. 
Valid values that correspond to regions used in NRSA are
CPL, NAP, NPL, SAP, sPL, TPL, UMW, WMT, and XER.}

\item{lwsarea}{Numeric values representing the log10 of 
watershed area in square km.}
}
\value{
A data frame containing the variables in sampID, watershed-
adjusted metrics, scored metrics, and the fish MMI for each site. 
Each of these components will be in separate columns, but values 
will only be provided for the metrics in the MMI for the specific 
region of each site, with all other columns being missing for 
that site.
}
\description{
This is a function that calculates 
the fish MMI as used for the National Rivers and Streams 
Assessment, based on inputs of the appropriate metrics.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
\keyword{survey}
