% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFishNativeMets.r
\name{calcFishNativeMets}
\alias{calcFishNativeMets}
\title{Calculate all native status metrics}
\usage{
calcFishNativeMets(indata, sampID = "UID", dist = "IS_DISTINCT",
  ct = "TOTAL", taxa_id = "TAXA_ID", nonnat = "NONNATIVE")
}
\arguments{
\item{indata}{Input data frame containing variables as identified
in the arguments for \emph{sampID}, \emph{dist}, \emph{ct}, 
\emph{taxa_id}, and \emph{nonnat}.}

\item{sampID}{sampID A character vector containing the names of all 
variables in indf that specify a unique sample. If not specified, 
the default is \emph{UID}.}

\item{dist}{A string with the name of the distinctness variable, 
which is assumed to have only values of 0 or 1. If not specified, 
the default is \emph{IS_DISTINCT}.}

\item{ct}{A string with the name of the count variable. If not 
specified, the default is \emph{TOTAL}.}

\item{taxa_id}{A string with the name of the taxon variable 
in \emph{indata}. The default value is \emph{TAXA_ID}.}

\item{nonnat}{A string with the name of the optional variable in 
\emph{inCts} containing non-native status. Valid values are 'Y' for 
non-native and 'N' for native. The default name 
is \emph{NONNATIVE}.}
}
\value{
A data frame containing the variables in sampID and 
the fish native status metrics as additional variables. Metric 
descriptions are included in \emph{NRSA_Fish_Metric_Descriptions.pdf},
included in this package. The names of
metrics include  NAT_TOTLNTAX, NAT_PTAX, NAT_PIND, ALIENNTAX, 
ALIENPTAX, ALIENPIND.
}
\description{
This function calculates all of the metrics based
only on native status, including alien metrics.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
