% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFishTaxMetSubsets.r
\name{calcFishTaxMets}
\alias{calcFishTaxMets}
\title{Calculate all taxonomy-related metrics}
\usage{
calcFishTaxMets(indata, inTaxa = NULL, sampID = "UID",
  dist = "IS_DISTINCT", ct = "TOTAL", taxa_id = "TAXA_ID",
  nonnat = "NONNATIVE", family = "FAMILY", genus = "GENUS",
  comname = "NAME")
}
\arguments{
\item{indata}{Input data frame containing variables as identified
in the arguments for \emph{sampID}, \emph{dist}, \emph{ct}, 
\emph{taxa_id}, as well as the optional variable 
for non-native status in \emph{nonnat}.}

\item{inTaxa}{Data frame containing fish taxalist, along with autecology
traits. At a minimum, this taxalist must contain variables matching 
the argument for \emph{family}, \emph{genus}, and \emph{comname}. 
The variable specified in the argument \emph{taxa_id} 
must match that in \emph{indata}.}

\item{sampID}{sampID A character vector containing the names of all 
variables in indf that specify a unique sample. If not specified, 
the default is \emph{UID}.}

\item{dist}{A string with the name of the distinctness variable, 
which is assumed to have only values of 0 or 1. If not specified, 
the default is \emph{IS_DISTINCT}.}

\item{ct}{A string with the name of the count variable. If not 
specified, the default is \emph{TOTAL}.}

\item{taxa_id}{A string with the name of the taxon ID variable 
in \emph{indata} that matches that in \emph{inTaxa}. The default 
value is \emph{TAXA_ID}.}

\item{nonnat}{A string with the name of the optional variable in 
\emph{inCts} containing non-native status. Valid values are 'Y' for 
non-native and 'N' for native. The default name 
is \emph{NONNATIVE}.}

\item{family}{A string with the name of the variable in the 
\emph{inTaxa} taxalist containing family name. The default value 
is \emph{FAMILY}.}

\item{genus}{A string with the name of the variable in the 
\emph{inTaxa} taxalist containing the genus name. The default
value is \emph{GENUS}}

\item{comname}{A string with the name of the variable in the 
\emph{inTaxa} taxalist containing the common name. The 
default value is \emph{NAME}.}
}
\value{
A data frame containing the variables in sampID and 
the fish tolerance metrics as additional variables. Metric 
descriptions are included in \emph{NRSA_Fish_Metric_Descriptions.pdf},
included in this package. The names of
metrics include  SALMPIND, SALMNTAX, SALMPTAX, ICTAPIND, ICTANTAX, 
ICTAPTAX, CATOPIND, CATONTAX, CATOPTAX, RBCATOPIND, RBCATONTAX, 
RBCATOPTAX, CENTPIND, CENTNTAX, CENTPTAX, CYPRPIND, CYPRNTAX, CYPRPTAX,
TOTLNTAX, and TOTLNIND. 

If a non-native status variable is included, these metrics are also
calculated:
NAT_SALMPIND, NAT_SALMNTAX, NAT_SALMPTAX, NAT_ICTAPIND, NAT_ICTANTAX, 
NAT_ICTAPTAX, NAT_CATOPIND, NAT_CATONTAX, NAT_CATOPTAX, NAT_RBCATOPIND, 
NAT_RBCATONTAX, NAT_RBCATOPTAX, NAT_CENTPIND, NAT_CENTNTAX, NAT_CENTPTAX, 
NAT_CYPRPIND, NAT_CYPRNTAX, NAT_CYPRPTAX, NAT_TOTLNTAX, NAT_TOTLNIND,
NAT_PIND, NAT_PTAX.
}
\description{
This function calculates all of the taxonomy metrics,
and if additional trait values are included in the taxalist (habitat,
trophic, migr). Native status versions of metrics are calculated
if a non-native status variable is included in the input count data.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
