% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFishTrophicMets.r
\name{calcFishTrophicMets}
\alias{calcFishTrophicMets}
\title{Calculate all trophic-related metrics}
\usage{
calcFishTrophicMets(indata, inTaxa = NULL, sampID = "UID",
  dist = "IS_DISTINCT", ct = "TOTAL", taxa_id = "TAXA_ID",
  habitat = "HABITAT", trophic = "TROPHIC", nonnat = "NONNATIVE")
}
\arguments{
\item{indata}{Input data frame containing variables as identified
in the arguments for \emph{sampID}, \emph{dist}, \emph{ct}, 
\emph{taxa_id}, as well as the optional variable 
for non-native status in \emph{nonnat}.}

\item{inTaxa}{Data frame containing fish taxalist, along with autecology
traits. At a minimum, this taxalist must contain variables matching 
the argument for \emph{trophic}. If habitat values are included, as 
specified by the argument \emph{habitat} additional metrics that combine 
trophic status and habitat are also calculated. The variable specified 
in the argument \emph{taxa_id} must match that in \emph{indata}.}

\item{sampID}{sampID A character vector containing the names of all 
variables in indf that specify a unique sample. If not specified, 
the default is \emph{UID}.}

\item{dist}{A string with the name of the distinctness variable, 
which is assumed to have only values of 0 or 1. If not specified, 
the default is \emph{IS_DISTINCT}.}

\item{ct}{A string with the name of the count variable. If not 
specified, the default is \emph{TOTAL}.}

\item{taxa_id}{A string with the name of the taxon ID variable 
in \emph{indata} that matches that in \emph{inTaxa}. The default 
value is \emph{TAXA_ID}.}

\item{habitat}{An optional string with the name of the variable in 
\emph{inTaxa} containing habitat preference values. Valid
values include B (benthic), W (water column), E (edge), or
blank, if unknown. The default value is \emph{HABITAT}.}

\item{trophic}{A string with the name of the variable in
\emph{inTaxa} containing trophic category values. Valid
values include I (invertivore), C (carnivore), O (omnivore),
H (herbivore), or blank if unknown. The default value is
\emph{inTaxa} is TROPHIC.}

\item{nonnat}{A string with the name of the optional variable in 
\emph{inCts} containing non-native status. Valid values are 'Y' for 
non-native and 'N' for native. The default name 
is \emph{NONNATIVE}.}
}
\value{
A data frame containing the variables in sampID and 
the fish trophic metrics as additional variables. Metric 
descriptions are included in \emph{NRSA_Fish_Metric_Descriptions.pdf},
included in this package. The names of
metrics include  INVNTAX, INVPIND, INVPTAX, CARNNTAX, CARNPIND, 
CARNPTAX, OMNINTAX, OMNIPIND, OMNIPTAX, HERBNTAX, HERBPIND, 
HERBPTAX, TOTLNIND, and TOTLNTAX.

If a non-native status variable is included, these metrics are also
calculated:
NAT_INVNTAX, NAT_INVPIND, NAT_INVPTAX, NAT_CARNNTAX, 
NAT_CARNPIND, NAT_CARNPTAX, NAT_OMNINTAX, NAT_OMNIPIND, 
NAT_OMNIPTAX, NAT_HERBNTAX, NAT_HERBPIND, NAT_HERBPTAX, 
NAT_TOTLNTAX, NAT_TOTLNIND, NAT_PIND, NAT_PTAX. 
 
Additional metrics calculated if habitat is included in 
\emph{inTaxa}: BENTINVNIND, BENTINVNTAX, BENTINVPIND. 

If a non-native status variable is included, these additional metrics
are also calculated: NAT_BENTINVNIND, NAT_BENTINVNTAX, NAT_BENTINVPIND.
}
\description{
This function calculates all of the trophic metrics,
and if habitat values are included in the taxalist, additional 
metrics are calculated.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
