% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNLA_BentMMImets.r
\name{calcNLA_BentMMImets}
\alias{calcNLA_BentMMImets}
\title{Calculate benthic MMI metrics used in NLA MMIs}
\usage{
calcNLA_BentMMImets(inCts, inTaxa = bentTaxa_nla, sampID = "UID",
  ecoreg = NULL, dist = "IS_DISTINCT", ct = "TOTAL",
  taxa_id = "TAXA_ID", ffg = "FFG", habit = "HABIT", ptv = "PTV")
}
\arguments{
\item{inCts}{A data frame containing, at minimum, the variables 
specified in the arguments for sampID, dist, ct, taxa_id, and
ecoreg. It is assumed that the data have been aggregated
to the taxonomic levels used in WSA/NRSA already. This can
be done using the function \emph{calcNLA_BentMMImets()}.}

\item{inTaxa}{a data frame containing taxonomic information, 
including variables for PHYLUM, CLASS, ORDER, FAMILY, SUBFAMILY, 
and TRIBE, as well as autecology traits with names that match those 
in the arguments ffg, habit, and ptv. In addition, there
should be a variable with the name in argument taxa_id that matches 
with all of those in the inCts data frame. The default is the
bentTaxa_nla data frame included in this package.}

\item{sampID}{A character vector containing the names of all 
variables in inCts that specify a unique sample. If not specified, 
the default is \emph{UID}}

\item{ecoreg}{A string with the name of the aggregated bioregion variable. 
Valid values that correspond to regions used in NLA are
CPL, EHIGH, PLAINS, UMW, and WMTNS.}

\item{dist}{A string with the name of the distinctness variable, 
which is assumed to have only values of 0 or 1. If not specified, 
the default is \emph{IS_DISTINCT}.}

\item{ct}{A string with the name of the count variable. If not 
specified, the default is \emph{TOTAL}.}

\item{taxa_id}{A string with the name of the taxon ID variable 
in \emph{inCts} that matches that in \emph{inTaxa}. The default 
value is \emph{TAXA_ID}.}

\item{ffg}{A string with the name of the functional feeding group 
variable in inTaxa. The default value is \emph{FFG}. Values used
in calculations include CF, CG, PR, SH, Sc, representing 
collector-filterer, collector-gatherer, predator, shredder, and
scraper, respectively. Each taxon may have more than 
one FFG value.}

\item{habit}{A string with the name of the habit variable in inTaxa. 
The default value is \emph{HABIT}. Values for habit that are used in
calculations include BU, CB, CN, SP, SW, representing burrower, 
climber, clinger, sprawler, and swimmer, respectively. Each taxon 
may have more than one value for HABIT.}

\item{ptv}{A string with the name of the pollution tolerance value 
variable in inTaxa. The default is \emph{PTV}.}
}
\value{
A data frame containing the variables in sampID and all of 
the benthic macroinvertebrate metrics used in the MMI as additional variables.
The metrics generated, by aggregated ecoregion, are:

 CPL: NOINPTAX, CHIRDOM5PIND, PREDNTAX, SPWLNTAX, EPT_NTAX, NTOLPIND
     
 EHIGH: NOINPTAX, CHIRDOM5PIND, COGANTAX, CLNGNTAX, EPOTNTAX, TL23NTAX
 
 PLAINS: DIPTPTAX, CHIRDOM5PIND, PREDNTAX, CLMBPTAX, EPOTNTAX, TL23PIND
 
 UMW: NOINPIND, CHIRDOM3PIND, SHRDPIND, CLNGNTAX, CRUSNTAX, TL23PTAX
 
 WMTNS: DIPTPIND, HPRIME, SCRPNTAX, CLNGNTAX, EPT_NTAX, TL23PTAX

Metric descriptions are included in \emph{NRSA_Fish_Metric_Descriptions.pdf},
included in this package.
}
\description{
This function calculates only the benthic
metrics in the corresponding MMI used in the National
Lakes Assessment (NLA), based on 
Omernik ecoregions aggregated to 5 bioregions included in the input 
data frame.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
\keyword{survey}
