% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNRSA_BenthicMMI.r
\name{calcNRSA_BenthicMMI}
\alias{calcNRSA_BenthicMMI}
\title{Calculate the NRSA benthic macroinvertebrate MMI}
\usage{
calcNRSA_BenthicMMI(inMets, sampID = "UID", ecoreg = "ECOREG",
  totlnind = "TOTLNIND")
}
\arguments{
\item{inMets}{A data frame containing, at minimum, the variables 
specified in the arguments for sampID, ecoreg, totlnind, and the metrics 
necessary for calculation of MMI by region. If values for more than 
the necessary metrics are included in the input data frame, the 
unnecessary  metrics will be ignored for each given site.

The necessary metrics, by aggregate ecoregion, are:

 CPL: NOINPIND, HPRIME, SHRDNTAX, CLNGPTAX, EPT_NTAX, TOLRPTAX
     
 NAP: EPT_PTAX, DOM5PIND, SCRPNTAX, CLNGPTAX, EPT_NTAX, NTOLPTAX
     
 NPL: EPT_PTAX, HPRIME, SCRPNTAX, BURRPTAX, EPHENTAX, NTOLNTAX 
     
 SAP: EPHEPTAX, HPRIME, SCRPNTAX, BURRPTAX, EPT_NTAX, TOLRPTAX
     
 SPL: EPT_PIND, HPRIME, SCRPNTAX, BURRPTAX, EPT_NTAX, INTLNTAX
 
 TPL: EPT_PIND, HPRIME, SCRPNTAX, CLNGNTAX, EPHENTAX, STOLPTAX
     
 UMW: CHIRPTAX, HPRIME, SHRDNTAX, BURRPTAX, EPT_NTAX, STOLPTAX
     
 WMT: EPT_PTAX, DOM5PIND, SCRPNTAX, CLNGPTAX, EPT_NTAX, TOLRPTAX
     
 XER: NOINPIND, DOM5PIND, SCRPNTAX, CLNGPTAX, EPT_NTAX, TOLRPTAX
 
 Descriptions of these metrics can be found in the file 
 \emph{NRSA_Invertebrate_Metric_Descriptions.pdf}, included in 
 the documentation for this package.}

\item{sampID}{A character vector containing the names of all 
variables in inMets that specify a unique sample. If not specified, 
the default is \emph{UID}}

\item{ecoreg}{A string with the name of the ecoregion variable. 
Valid values that correspond to regions used in NRSA are
CPL, NAP, NPL, SAP, sPL, TPL, UMW, WMT, and XER.}

\item{totlnind}{A string with the name of the variable with the 
total individuals in each sample.}
}
\value{
A data frame containing the variables in sampID, as well as
the scored metrics, the benthic MMI, and the condition class for each
sites. The variable names are COMP_PT, DIVS_PT, FEED_PT, HABT_PT,
RICH_PT, TOLR_PT, MMI_BENT, BENT_MMI_COND.
}
\description{
This is a function that calculates 
the benthic MMI as used for the National Rivers and Streams 
Assessment, based on inputs of the appropriate metrics.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
\keyword{survey}
