% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNRSA_FishMMImets.r
\name{calcNRSA_FishMMImets}
\alias{calcNRSA_FishMMImets}
\title{Calculate only fish metrics necessary for NRSA MMI}
\usage{
calcNRSA_FishMMImets(indata, inTaxa = NULL, sampID = "UID", ecoreg = NULL,
  dist = "IS_DISTINCT", ct = "TOTAL", taxa_id = "TAXA_ID",
  tol = "TOLERANCE", vel = "VELOCITY", habitat = "HABITAT",
  trophic = "TROPHIC", migr = "MIGRATORY", nonnat = "NONNATIVE",
  reprod = "REPROD", family = "FAMILY", genus = "GENUS",
  comname = "NAME")
}
\arguments{
\item{indata}{A data frame containing, at minimum, the variables 
specified in the arguments for \emph{sampID}, \emph{dist}, 
\emph{ct}, and \emph{taxa_id}, as well as the optional variable 
for non-native status in \emph{nonnat}.}

\item{inTaxa}{a data frame containing taxonomic information, 
including variables that match the arguments for (\emph{family}), 
(\emph{genus}), and (\emph{comname}), as well as autecology traits 
with names that match those in the arguments \emph{tol}, \emph{vel}, 
\emph{habitat}, \emph{trophic}, \emph{migr}, and \emph{reprod}. 
In addition, there should be a variable with the 
name in argument \emph{taxa_id} that matches 
with all of those in the indata data frame}

\item{sampID}{A character vector containing the names of all 
variables in indata that specify a unique sample. If not specified, 
the default is \emph{UID}}

\item{ecoreg}{A string with the name of the ecoregion variable. 
Valid values that correspond to regions used in NRSA are
CPL, NAP, NPL, SAP, sPL, TPL, UMW, WMT, and XER.}

\item{dist}{A string with the name of the distinctness variable, 
which is assumed to have only values of 0 or 1. If not specified, 
the default is \emph{IS_DISTINCT}.}

\item{ct}{A string with the name of the count variable. If not 
specified, the default is \emph{FINAL_CT}.}

\item{taxa_id}{A string with the name of the taxon ID variable 
in \emph{indata} that matches that in \emph{inTaxa}. The default 
value is \emph{TAXA_ID}.}

\item{tol}{A string with the name of the variable in the 
\emph{inTaxa} taxalist containing tolerance categories. Valid
values include S (sensitive), I (intermediate), and T (tolerant).
The default value is \emph{TOLERANCE}.}

\item{vel}{A strings with the name of the variable in the 
\emph{inTaxa} taxalist containing velocity preference values.
Valid values include R (Rheophil), P (Pool), O (Other), or 
blank, if unknown. The default value is \emph{VELOCITY}.}

\item{habitat}{A string with the name of the variable in 
\emph{inTaxa} containing habitat preference values. Valid
values include B (benthic), W (water column), E (edge), or
blank, if unknown. The default value is \emph{HABITAT}.}

\item{trophic}{A string with the name of the variable in
\emph{inTaxa} containing trophic category values. Valid
values include I (invertivore), C (carnivore), O (omnivore),
H (herbivore), or blank if unknown. The default value is
\emph{inTaxa} is TROPHIC.}

\item{migr}{A string with the name of the variable in 
\emph{inTaxa} containing migratory status value. Valid 
values include N (No), Y (Yes), and blank if unknown. The
default value is \emph{MIGRATORY}.}

\item{nonnat}{A string with the name of the optional variable in 
\emph{inCts} containing non-native status. Valid values are 'Y' for 
non-native and 'N' for native. The default name 
is \emph{NONNATIVE}.}

\item{reprod}{A string with the name of the variable in 
\emph{inTaxa} containing reproductive trait values. Valid 
values include C (Clean, coarse lithophil), D (Drifter),
G (Guarder), O (Other), or blank if unknown. The default
value is \emph{REPROD}.}

\item{family}{A string with the name of the variable in the 
\emph{inTaxa} taxalist containing family name. The default value 
is \emph{FAMILY}.}

\item{genus}{A string with the name of the variable in the 
\emph{inTaxa} taxalist containing the genus name. The default
value is \emph{GENUS}}

\item{comname}{A string with the name of the variable in the 
\emph{inTaxa} taxalist containing the common name. The 
default value is \emph{NAME}.}
}
\value{
A data frame containing the variables in sampID, 
the total number of individuals in the sample as TOTLNIND, and 
all fish metrics used in MMIs as additional variables. All metrics
will be separate columns, but values will only be provided for 
the metrics in the MMI for the specific region of each site,
with all other metrics being missing for that site. The metrics, 
by aggregated ecoregion, are:

 CPL: ALIENPIND, RBCATONTAX, LOTPIND, INTLMIGRPTAX, LITHPIND, NAT_TOTLNTAX, 
     TOLRNTAX, INVPTAX
     
 NAP: ALIENNTAX, SALMNTAX, NAT_RHEOPIND, INTLMIGRPIND, LITHPTAX,
     NTOLPTAX, TOLRNTAX, INVNTAX
     
 NPL: ALIENNTAX, NAT_CYPRPIND, LOTNTAX, MIGRNTAX, LITHPIND, NTOLPTAX,
     NAT_INTLPIND, NAT_CARNNTAX 
     
 SAP: NAT_PTAX, NAT_CENTNTAX, NAT_NTOLBENTPTAX, NAT_MIGRNTAX,
     NAT_LITHPIND, NTOLPTAX, TOLRPTAX, INVPIND
     
 SPL: NAT_PIND, CYPRPTAX, RHEOPIND, NAT_MIGRPTAX, LITHNTAX,
     NAT_NTOLNTAX, TOLRNTAX, HERBPTAX
 
 TPL: ALIENNTAX, NAT_ICTAPIND, RHEONTAX, INTLMIGRNTAX, LITHPIND,
     NAT_NTOLNTAX, INTLPTAX, CARNNTAX
     
 UMW: NAT_PTAX, CYPRNTAX, INTLLOTNTAX, INTLMIGRPTAX, LITHPIND,
     NTOLNTAX, TOLRNTAX, INTLINVPTAX
     
 WMT: NAT_PIND, NAT_CATOPIND, INTLLOTPTAX, NAT_MIGRPTAX, LITHPTAX,
     NAT_TOTLNTAX, TOLRNTAX, NAT_HERBPTAX
     
 XER: NAT_PIND, CENTPTAX, RHEOPIND, MIGRPTAX, LITHNTAX, NTOLPTAX,
     TOLRNTAX, BENTINVPTAX

Metric descriptions are included in \emph{NRSA_Fish_Metric_Descriptions.pdf},
included in this package.
}
\description{
This is a function that calculates 
only the fish metrics used in the National Rivers and Streams
Assessment fish MMI.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
\keyword{survey}
