% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLAphab_indicators.r
\name{nlaDrawdownIndicator}
\alias{nlaDrawdownIndicator}
\title{Calculate NLA Drawdown Indicator}
\usage{
nlaDrawdownIndicator(x, sampID, bfxVertDD, bfxHorizDD, ecoreg, lake_origin,
  bfnHorizDD_nomod = NULL, bfnVertDD = NULL)
}
\arguments{
\item{x}{Input data frame containing other variables as specified in
arguments}

\item{sampID}{Vector of variables necessary to identify a unique sample 
in \emph{x}.}

\item{bfxVertDD}{Vertical height to highwater mark in meters}

\item{bfxHorizDD}{Horizontal distance to highwater mark in meters}

\item{ecoreg}{Lake ecoregion, based on aggregated Omernick ecoregions,
with valid values of CPL, NAP, NPL, SAP, SPL, TPL, UMW, WMT, XER.}

\item{lake_origin}{Lake origin, with valid values of 'NATURAL' 
or 'MAN_MADE'}

\item{bfnHorizDD_nomod}{For 2017 data, number of unmodified horizontal 
distances measurements to highwater mark. Must be used with
\emph{bfnVertDD}}

\item{bfnVertDD}{For 2017 data, number of vertical distances measured
to highwater mark. Must be used with \emph{bfnHorizDD_nomod}}
}
\value{
A data frame containing:
\itemize{
\item{sampID}{The variables in the argument \emph{sampID}}

\item{horizDD_cond}{Drawdown condition based only on horizontal
distance}

\item{vertDD_cond}{Drawdown conditoin based only on vertical
height}

\item{DRAWDOWN_COND}{Drawdown condition class based on 
\emph{horizDD_cond} and \emph{vertDD_cond} values
 (Small/Medium/Large/Not Assessed)} 
 
\item{horizDD_cond17}{Drawdown condition based only on horizontal
distance using additional information on number of observations. 
Intended for 2017 data only. Run only if \emph{bfnHorizDD_nomod}
and \emph{bfnVertDD} arguments used.}

\item{vertDD_cond17}{Drawdown condition based only on vertical
height using additional information on number of observations. 
Intended for 2017 data only. Run only if \emph{bfnHorizDD_nomod}
and \emph{bfnVertDD} arguments used.}

\item{DDcond_screen}{Intermediate drawdown condition value based on
values of horizDD_cond17 and vertDD_cond17. Intended for 2017 data
only. Run only if \emph{bfnHorizDD_nomod}
and \emph{bfnVertDD} arguments used.}

\item{DRAWDOWN_COND_2017}{Drawdown condition value based on
values of horizDD_cond17, vertDD_cond17, and DDcond_screen. 
Intended for 2017 data
only. Run only if \emph{bfnHorizDD_nomod}
and \emph{bfnVertDD} arguments used. (Small/Medium/Large/Not Assessed)}

\item{DRAWDOWN_COND_2CAT_2017}{Drawdown condition value based on
combining Small and Medium values of DRAWDOWN_COND into "Not Large". 
Intended for 2017 data
only. Run only if \emph{bfnHorizDD_nomod}
and \emph{bfnVertDD} arguments used. (Not Large/Large/Not Assessed)}
}
}
\description{
Using metric values as inputs, calculate 
indicator score for lake drawdown, based on both horizontal
and vertical drawdown. Each is treated separately, and the 
two condition classes are then assessed to assign the 
larger of the two condition classes.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
\keyword{survey}
