% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlaHumanImpact.r
\name{nlaHumanImpact}
\alias{nlaHumanImpact}
\title{Calculate NLA Human Influence Metrics}
\usage{
nlaHumanImpact(buildings = NULL, buildings_dd = NULL, commercial = NULL,
  commercial_dd = NULL, crops = NULL, crops_dd = NULL, docks = NULL,
  docks_dd = NULL, landfill = NULL, landfill_dd = NULL, lawn = NULL,
  lawn_dd = NULL, orchard = NULL, orchard_dd = NULL, other = NULL,
  other_dd = NULL, park = NULL, park_dd = NULL, pasture = NULL,
  pasture_dd = NULL, powerlines = NULL, powerlines_dd = NULL,
  roads = NULL, roads_dd = NULL, walls = NULL, walls_dd = NULL,
  drawdown = NULL, horizontalDistance_dd = NULL, data2007 = FALSE,
  fillinDrawdown = TRUE, proximityWeights = data.frame(proximity = c("0",
  "P", "C"), calc = c(0, 0.5, 1), circa = c(0, 0, 1), present = c(0, 1, 1),
  stringsAsFactors = FALSE))
}
\arguments{
\item{buildings}{A data frame containing buildings human influence 
 proximity class values from riparian zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{buildings_dd}{A data frame containing buildings human influence 
 proximity class values from drawdown zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{commercial_dd}{A data frame containing commercial human influence 
 proximity class values from drawdown zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{crops}{A data frame containing row crops human influence 
 proximity class values from riparian zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{crops_dd}{A data frame containing row crops human influence 
 proximity class values from drawdown zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{docks}{A data frame containing docks or boats human influence 
 proximity class values from riparian zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{docks_dd}{A data frame containing docks or boats human influence 
 proximity class values from drawdown zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{landfill}{A data frame containing landfill/trash human influence 
 proximity class values from riparian zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{landfill_dd}{A data frame containing landfill/trash human influence 
 proximity class values from drawdown zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{lawn}{A data frame containing lawn human influence 
 proximity class values from riparian zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{lawn_dd}{A data frame containing lawn human influence 
 proximity class values from drawdown zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{orchard}{A data frame containing orchard human influence 
 proximity class values from riparian zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{orchard_dd}{A data frame containing orchard human influence 
 proximity class values from drawdown zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{other}{A data frame containing other human influence 
 proximity class values from riparian zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{other_dd}{A data frame containing other human influence 
 proximity class values from drawdown zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{park}{A data frame containing park facilities/man-made beach human 
influence proximity class values from riparian zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{park_dd}{A data frame containing park facilities/man-made beach 
human influence proximity class values from drawdown zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{pasture}{A data frame containing pasture/range/hay field human 
influence proximity class values from riparian zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{pasture_dd}{A data frame containing pasture/range/hay field human 
influence proximity class values from drawdown zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{powerlines}{A data frame containing power lines human influence 
proximity class values from riparian zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{powerlines_dd}{A data frame containing power lines human influence 
proximity class values from drawdown zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{roads}{A data frame containing roads or railroad human influence 
 proximity class values from riparian zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{roads_dd}{A data frame containing roads or railroad human influence 
 proximity class values from drawdown zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{walls}{A data frame containing walls, dikes, or revetments human 
influence proximity class values from riparian zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{walls_dd}{A data frame containing walls, dikes, or revetments 
human influence proximity class values from drawdown zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}

\item{drawdown}{A data frame containing the presence of drawdown at stations. 
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE an integer value, or character value that is castable to an 
integer, indicating the presence of drawdown at a station.
}}

\item{horizontalDistance_dd}{A data frame containing the horizontal 
distance from waterline to high water mark (m) when drawdown is present, 
with the columns:
\itemize{
\item SITE an integer or character value identifying a single site visit.
\item STATION a character value identifying the station within the SITE
\item VALUE an numeric value, or character value that is castable to an 
numeric.
}}

\item{data2007}{A logical value, which equals TRUE if 2007 data is being
processed.  The default value is FALSE.}

\item{fillinDrawdown}{A logical value, which specifies whether to use the
DRAWDOWN parameter to fill in unrecorded cover and horizontalDistance_dd 
values. The default value is TRUE.}

\item{proximityWeights}{A data frame relating categorical proximity values 
to various numeric weights for different types of metrics. The default data frame
consists of the following values:
\itemize{
\item proximity c('0','P','C')
\item calc c(0.0, 0.5, 1.0)  
\item circa c(0, 0, 1)
\item present c(0, 1, 1)
}}

\item{commerical}{A data frame containing commercial human influence 
 proximity class values from riparian zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of 0, P, or C representing the proximity 
category.
}}
}
\value{
Either a data frame when metric calculation is successful or a 
character string containing an error message when metric calculation 
is not successful. The data frame contains the following columns:
\itemize{ 
    \item SITE - unique site visit identifier
    \item METRIC - metric name
    \item VALUE - metric value
      }
The output metrics include:
 HIFPANY_DD, HIFPANY_RIP, HIFPANY_SYN, HIFPANYCIRCA_DD, 
 HIFPANYCIRCA_RIP, HIFPANYCIRCA_SYN, HIIAG_DD, HIIAG_RIP, 
 HIIAG_SYN, HIIAGCIRCA_DD, HIIAGCIRCA_RIP, HIIAGCIRCA_SYN, 
 HIIALL_DD, HIIALL_RIP, HIIALL_SYN, HIIALLCIRCA_DD, 
 HIIALLCIRCA_RIP, HIIALLCIRCA_SYN, HIINONAG_DD, HIINONAG_RIP, 
 HIINONAG_SYN, HIINONAGCIRCA_DD, HIINONAGCIRCA_RIP, HIINONAGCIRCA_SYN, 
 HINAG_DD, HINAG_RIP, HINAG_SYN, HINALL_DD, 
 HINALL_RIP, HINALL_SYN, HINBUILDINGS_DD, HINBUILDINGS_RIP, 
 HINBUILDINGS_SYN, HINCOMMERCIAL_DD, HINCOMMERCIAL_RIP, HINCOMMERCIAL_SYN, 
 HINCROPS_DD, HINCROPS_RIP, HINCROPS_SYN, HINDOCKS_DD, 
 HINDOCKS_RIP, HINDOCKS_SYN, HINLANDFILL_DD, HINLANDFILL_RIP, 
 HINLANDFILL_SYN, HINLAWN_DD, HINLAWN_RIP, HINLAWN_SYN, 
 HINNONAG_DD, HINNONAG_RIP, HINNONAG_SYN, HINORCHARD_DD, 
 HINORCHARD_RIP, HINORCHARD_SYN, HINOTHER_DD, HINOTHER_RIP, 
 HINOTHER_SYN, HINPARK_DD, HINPARK_RIP, HINPARK_SYN, 
 HINPASTURE_DD, HINPASTURE_RIP, HINPASTURE_SYN, HINPOWERLINES_DD, 
 HINPOWERLINES_RIP, HINPOWERLINES_SYN, HINROADS_DD, HINROADS_RIP, 
 HINROADS_SYN, HINWALLS_DD, HINWALLS_RIP, HINWALLS_SYN, 
 HIPWAG_DD, HIPWAG_RIP, HIPWAG_SYN, HIPWALL_DD, 
 HIPWALL_RIP, HIPWALL_SYN, HIPWBUILDINGS_DD, HIPWBUILDINGS_RIP, 
 HIPWBUILDINGS_SYN, HIPWCOMMERCIAL_DD, HIPWCOMMERCIAL_RIP, HIPWCOMMERCIAL_SYN
 HIPWCROPS_DD, HIPWCROPS_RIP, HIPWCROPS_SYN, HIPWDOCKS_DD, 
 HIPWDOCKS_RIP, HIPWDOCKS_SYN, HIPWLANDFILL_DD, HIPWLANDFILL_RIP, 
 HIPWLANDFILL_SYN, HIPWLAWN_DD, HIPWLAWN_RIP, HIPWLAWN_SYN, 
 HIPWNONAG_DD, HIPWNONAG_RIP, HIPWNONAG_SYN, HIPWORCHARD_DD, 
 HIPWORCHARD_RIP, HIPWORCHARD_SYN, HIPWOTHER_DD, HIPWOTHER_RIP, 
 HIPWOTHER_SYN, HIPWPARK_DD, HIPWPARK_RIP, HIPWPARK_SYN, 
 HIPWPASTURE_DD, HIPWPASTURE_RIP, HIPWPASTURE_SYN, HIPWPOWERLINES_DD, 
 HIPWPOWERLINES_RIP, HIPWPOWERLINES_SYN, HIPWROADS_DD, HIPWROADS_RIP, 
 HIPWROADS_SYN, HIPWWALLS_DD, HIPWWALLS_RIP, HIPWWALLS_SYN
 
Descriptions for all metrics are included in 
\emph{NLA_Physical_Habitat_Metric_Descriptions.pdf} in the package
documentation.
}
\description{
This function calculates the human influence 
portion of the physical habitat metrics for National Lakes Assessment 
(NLA) data.
}
\examples{
  head(nlaPhabEx)
  
  buildings <- subset(nlaPhabEx,PARAMETER=='HI_BUILDINGS',select=-PARAMETER)
  buildings_dd <- subset(nlaPhabEx,PARAMETER=='HI_BUILDINGS_DD',select=-PARAMETER)
  commercial <- subset(nlaPhabEx,PARAMETER=='HI_COMMERCIAL',select=-PARAMETER)
  commercial_dd <- subset(nlaPhabEx,PARAMETER=='HI_COMMERCIAL_DD',select=-PARAMETER)
  crops <- subset(nlaPhabEx,PARAMETER=='HI_CROPS',select=-PARAMETER)
  crops_dd <- subset(nlaPhabEx,PARAMETER=='HI_CROPS_DD',select=-PARAMETER)
  docks <- subset(nlaPhabEx,PARAMETER=='HI_DOCKS',select=-PARAMETER)
  docks_dd <- subset(nlaPhabEx,PARAMETER=='HI_DOCKS_DD',select=-PARAMETER)
  landfill <- subset(nlaPhabEx,PARAMETER=='HI_LANDFILL',select=-PARAMETER)
  landfill_dd <- subset(nlaPhabEx,PARAMETER=='HI_LANDFILL_DD',select=-PARAMETER)
  lawn <- subset(nlaPhabEx,PARAMETER=='HI_LAWN',select=-PARAMETER)
  lawn_dd <- subset(nlaPhabEx,PARAMETER=='HI_LAWN_DD',select=-PARAMETER)
  orchard <- subset(nlaPhabEx,PARAMETER=='HI_ORCHARD',select=-PARAMETER)
  orchard_dd <- subset(nlaPhabEx,PARAMETER=='HI_ORCHARD_DD',select=-PARAMETER)
  other <- subset(nlaPhabEx,PARAMETER=='HI_OTHER',select=-PARAMETER)
  other_dd <- subset(nlaPhabEx,PARAMETER=='HI_OTHER_DD',select=-PARAMETER)
  park <- subset(nlaPhabEx,PARAMETER=='HI_PARK',select=-PARAMETER)
  park_dd <- subset(nlaPhabEx,PARAMETER=='HI_PARK_DD',select=-PARAMETER)
  pasture <- subset(nlaPhabEx,PARAMETER=='HI_PASTURE',select=-PARAMETER)
  pasture_dd <- subset(nlaPhabEx,PARAMETER=='HI_PASTURE_DD',select=-PARAMETER)
  powerlines <- subset(nlaPhabEx,PARAMETER=='HI_POWERLINES',select=-PARAMETER)
  powerlines_dd <- subset(nlaPhabEx,PARAMETER=='HI_POWERLINES_DD',select=-PARAMETER)
  roads <- subset(nlaPhabEx,PARAMETER=='HI_ROADS',select=-PARAMETER)
  roads_dd <- subset(nlaPhabEx,PARAMETER=='HI_ROADS_DD',select=-PARAMETER)
  walls <- subset(nlaPhabEx,PARAMETER=='HI_WALLS',select=-PARAMETER)
  walls_dd <- subset(nlaPhabEx,PARAMETER=='HI_WALLS_DD',select=-PARAMETER)
  drawdown <- subset(nlaPhabEx,PARAMETER=='DRAWDOWN',select=-PARAMETER)
  horizontalDistance_dd <- subset(nlaPhabEx,PARAMETER=='HORIZ_DIST_DD',select=-PARAMETER)
  
  # Use defaults for data2007, fillinDrawdown, and proximityWeights
  # arguments
  exHumInfl <- nlaHumanImpact(buildings, buildings_dd, commercial, commercial_dd,
  crops, crops_dd,docks, docks_dd, landfill, landfill_dd, lawn, lawn_dd, orchard,
  orchard_dd, other, other_dd, park, park_dd, pasture, pasture_dd, powerlines,
  powerlines_dd, roads, roads_dd, walls, walls_dd, drawdown, horizontalDistance_dd)
  
  head(exHumInfl)
 
}
\author{
Curt Seeliger \email{Seeliger.Curt@epa.gov}\cr
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
