% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLAphab_indicators.r
\name{nlaLitRipVegCompIndicator}
\alias{nlaLitRipVegCompIndicator}
\title{Calculate NLA Littoral-riparian Vegetation Complexity Indicator}
\usage{
nlaLitRipVegCompIndicator(x, sampID, lat, lon, lake_origin, area, elev, ecoreg,
  rvegq, litcvrq)
}
\arguments{
\item{x}{Input data frame containing other variables as specified in
arguments}

\item{sampID}{Vector of variables necessary to identify a unique sample 
in \emph{x}.}

\item{lat}{Latitude of lake in decimal degrees (NAD 83 datum)}

\item{lon}{Longitude of lake in decimal degrees (NAD 83 datum)}

\item{lake_origin}{Lake origin, with valid values of 'NATURAL' 
or 'MAN_MADE'}

\item{area}{Area of lake or reservoir, in square kilometers}

\item{elev}{Lake surface elevation (meters above sea level)}

\item{ecoreg}{Lake ecoregion, based on aggregated Omernick ecoregions,
with valid values of CPL, NAP, NPL, SAP, SPL, TPL, UMW, WMT, XER.}

\item{rvegq}{Observed riparian vegetation complexity index value, 
calculated by function \code{nlaRipVegCompIndicator()}.}

\item{litcvrq}{Observed littoral vegetation complexity index value,
calculated by function \code{nlaLitVegCompIndicator()}.}
}
\value{
A data frame containing:
\itemize{
\item{sampID}{The variables in the argument \emph{sampID}}

\item{LitRipCvrQ}{Observed combined riparian and littoral vegetation
complexity index}

\item{LitRipCvrQc3x15}{Expected combined riparian and littoral
vegetation complexity index}

\item{LitRipCvrQc3OE}{Littoral and riparian vegetation complexity 
Observed/Expected (O/E) value}

\item{LITRIPCVR_COND}{Littoral and riparian vegetation complexity 
indicator condition class (Good/Fair/Poor/Not Assessed)} 
}
}
\description{
Using metric values and various predictors as inputs, calculate 
indicator score for LitRipCvrQc3OE, the observed over expected value for
combined littoral and riparian vegetation complexity. Models to calculate 
expected values depend on ecoregion. Use thresholds based on reference 
distributions by ecoregion to assign condition.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
\keyword{survey}
