% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLAphab_indicators.r
\name{nlaLitVegCompIndicator}
\alias{nlaLitVegCompIndicator}
\title{Calculate NLA Littoral Vegetation Complexity Indicator}
\usage{
nlaLitVegCompIndicator(x, sampID, lat, lon, lake_origin, area, elev, ecoreg,
  fciNatural, fcfcSnag, amfcFloating, amfcEmergent, fcfcBoulders, fcfcBrush,
  fcfcLedges, fcfcLiveTrees, fcfcOverhang)
}
\arguments{
\item{x}{Input data frame containing other variables as specified in
arguments}

\item{sampID}{Vector of variables necessary to identify a unique sample 
in \emph{x}.}

\item{lat}{Latitude of lake in decimal degrees (NAD 83 datum)}

\item{lon}{Longitude of lake in decimal degrees (NAD 83 datum)}

\item{lake_origin}{Lake origin, with valid values of 'NATURAL' 
or 'MAN_MADE'}

\item{area}{Area of lake or reservoir, in square kilometers}

\item{elev}{Lake surface elevation (meters above sea level)}

\item{ecoreg}{Lake ecoregion, based on aggregated Omernick ecoregions,
with valid values of CPL, NAP, NPL, SAP, SPL, TPL, UMW, WMT, XER.}

\item{fciNatural}{Index of fish cover due to natural structures.
Calculated by the function \code{metsFishCoverNLA()}.}

\item{fcfcSnag}{Fractional fish cover consisting of snags.
Calculated by the function \code{metsFishCoverNLA()}.}

\item{amfcFloating}{Fractional cover of floating aquatic macrophytes.
Calculated by the function \code{metsFishCoverNLA()}.}

\item{amfcEmergent}{Fractional cover of emergent aquatic macrophytes.
Calculated by the function \code{metsFishCoverNLA()}.}

\item{fcfcBoulders}{Fraction of fish cover as boulders in 
littoral zone. Calculated by the function \code{metsFishCoverNLA()}.}

\item{fcfcBrush}{Fraction of fish cover as brush. 
Calculated by the function \code{metsFishCoverNLA()}.}

\item{fcfcLedges}{Fraction of fish cover as ledges.
Calculated by the function \code{metsFishCoverNLA()}.}

\item{fcfcLiveTrees}{Fractino of fish cover as live trees.
Calculated by the function \code{metsFishCoverNLA()}.}

\item{fcfcOverhang}{Fraction of fish cover as overhangs.
Calculated by the function \code{metsFishCoverNLA()}.}
}
\value{
A data frame containing:
\itemize{
\item{sampID}{The variables in the argument \emph{sampID}}

\item{LitCvrQ}{Observed littoral vegetation complexity index}

\item{LitCvrQc3x15}{Expected littoral vegetation complexity index}

\item{LitCvrQc3OE}{Littoral vegetation complexity indicator 
value, O/E score}

\item{LITCVR_COND}{Littoral vegetation complexity indicator 
condition class (Good/Fair/Poor/Not Assessed)} 
}
}
\description{
Using metric values and various predictors as inputs, calculate 
indicator score for LitCvrQc3OE, the observed over expected value for
littoral vegetation complexity. Models to calculate expected values depend
on ecoregion. Use thresholds based on reference distributions by ecoregion 
to assign condition.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
\keyword{survey}
