% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlaPhab}
\alias{nlaPhab}
\title{Example data for NLA physical habitat field measurements}
\format{
a data frame with 7449 rows and 4 variables:
\describe{
 \item{SITE}{int Site visit identifier}
 \item{STATION}{chr Station, range from A-I}
 \item{PARAMETER}{chr Parameter name for field measurement}
 \item{VALUE}{chr Value of parameter, combination of numeric and string
 values}
}
}
\source{
aquamet package
}
\usage{
nlaPhab
}
\description{
Example data for NLA physical habitat field measurements
}
\keyword{datasets}
