\name{nlaPhabMetsEx}
\alias{nlaPhabMetsEx}
\docType{data}
\title{Example NLA Physical Habitat Metrics Dataset}
\description{
  A dataset containing NLA physical habitat metric data for use in function examples.
}
\usage{data(nlaPhabMetsEx)}
\format{
  A data frame with 5114 observations on the following 3 variables.
  \describe{
    \item{SITE}{unique site visit ID.}
    \item{METRIC}{character variable of calculated metric.}
    \item{VALUE}{value of metric.}
  }
}
\details{
  These data are a small subset of the NLA physical habitat metrics dataset for example purposes only.
}
\examples{
  data(nlaPhabMetsEx)
  head(nlaPhabMetsEx)
}
\keyword{datasets}
