% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlaRiparianVegetation.r
\name{nlaRiparianVegetation}
\alias{nlaRiparianVegetation}
\title{Calculate NLA Riparian Zone and Vegetation Metrics}
\usage{
nlaRiparianVegetation(bigTrees = NULL, bigTrees_dd = NULL,
  smallTrees = NULL, smallTrees_dd = NULL, canopyType = NULL,
  canopyType_dd = NULL, groundcoverBare = NULL, groundcoverBare_dd = NULL,
  groundcoverInundated = NULL, groundcoverInundated_dd = NULL,
  groundcoverNonwoody = NULL, groundcoverNonwoody_dd = NULL,
  groundcoverWoody = NULL, groundcoverWoody_dd = NULL,
  understoryNonwoody = NULL, understoryNonwoody_dd = NULL,
  understoryWoody = NULL, understoryWoody_dd = NULL,
  understoryType = NULL, understoryType_dd = NULL, drawdown = NULL,
  horizontalDistance_dd = NULL, createSyntheticCovers = TRUE,
  fillinDrawdown = TRUE)
}
\arguments{
\item{bigTrees}{A data frame containing cover class values for 
big trees (trunk >0.3 m dBH) in the riparian zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE an integer value, or character value that is castable to an 
integer, from 0-4 containing the cover category.
}}

\item{bigTrees_dd}{A data frame containing cover class values for 
big trees (trunk >0.3 m dBH) in the drawdown zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE an integer value, or character value that is castable to an 
integer, from 0-4 containing the cover category.
}}

\item{smallTrees}{A data frame containing cover class values for 
small trees (trunk <0.3 m dBH) in the riparian zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE an integer value, or character value that is castable to an 
integer, from 0-4 containing the cover category.
}}

\item{smallTrees_dd}{A data frame containing cover class values for 
small trees (trunk <0.3 m dBH) in the drawdown zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE an integer value, or character value that is castable to an 
integer, from 0-4 containing the cover category.
}}

\item{canopyType}{A data frame containing canopy type (> 5 m) values for 
in the riparian zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of C (coniferous), D (deciduous), E 
(broadleaf evergreen), M (mixed), or N (None) the cover type.
}}

\item{canopyType_dd}{A data frame containing canopy type (> 5 m) values 
for in the drawdown zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of C (coniferous), D (deciduous), E 
(broadleaf evergreen), M (mixed), or N (None) the cover type.
}}

\item{groundcoverBare}{A data frame containing cover class values for 
barren, bare dirt, litter, duff, or building ground cover (< 0.5 m) 
types in the riparian zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE an integer value, or character value that is castable to an 
integer, from 0-4 containing the cover category.
}}

\item{groundcoverBare_dd}{A data frame containing cover class values for 
barren, bare dirt, litter, duff, or building ground cover (< 0.5 m) 
types in the drawdown zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE an integer value, or character value that is castable to an 
integer, from 0-4 containing the cover category.
}}

\item{groundcoverInundated}{A data frame containing cover class values for 
standing water or inundated vegetation ground cover (< 0.5 m) types in the 
riparian zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE an integer value, or character value that is castable to an 
integer, from 0-4 containing the cover category.
}}

\item{groundcoverInundated_dd}{A data frame containing cover class values for 
standing water or inundated vegetation ground cover (< 0.5 m)types in the 
drawdown zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE an integer value, or character value that is castable to an 
integer, from 0-4 containing the cover category.
}}

\item{groundcoverNonwoody}{A data frame containing cover class values for 
herbs, grasses, and forbs ground cover (< 0.5 m) types in the 
riparian zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE an integer value, or character value that is castable to an 
integer, from 0-4 containing the cover category.
}}

\item{groundcoverNonwoody_dd}{A data frame containing cover class values for 
herbs, grasses, and forbs ground cover (< 0.5 m) types in the 
drawdown zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE an integer value, or character value that is castable to an 
integer, from 0-4 containing the cover category.
}}

\item{groundcoverWoody}{A data frame containing cover class values for 
woody shrubs and saplings ground cover (< 0.5 m) types in the 
riparian zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE an integer value, or character value that is castable to an 
integer, from 0-4 containing the cover category.
}}

\item{groundcoverWoody_dd}{A data frame containing cover class values for 
woody shrubs and saplings ground cover (< 0.5 m) types in the 
drawdown zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE an integer value, or character value that is castable to an 
integer, from 0-4 containing the cover category.
}}

\item{understoryNonwoody}{A data frame containing cover class values for 
tall herbs, grasses, and forbs understory cover (0.5 to 5 m) types in the 
riparian zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE an integer value, or character value that is castable to an 
integer, from 0-4 containing the cover category.
}}

\item{understoryNonwoody_dd}{A data frame containing cover class values for 
tall herbs, grasses, and forbs understory cover (0.5 to 5 m) types in the 
drawdown zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE an integer value, or character value that is castable to an 
integer, from 0-4 containing the cover category.
}}

\item{understoryWoody}{A data frame containing cover class values for 
woody shrubs and saplings understory cover (0.5 to 5 m) types in the 
riparian zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE an integer value, or character value that is castable to an 
integer, from 0-4 containing the cover category.
}}

\item{understoryWoody_dd}{A data frame containing cover class values for 
woody shrubs and saplings understory cover (0.5 to 5 m) types in the 
drawdown zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE an integer value, or character value that is castable to an 
integer, from 0-4 containing the cover category.
}}

\item{understoryType}{A data frame containing understory type (0.5 to 5 m) 
values for in the riparian zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of C (coniferous), D (deciduous), E 
(broadleaf evergreen), M (mixed), or N (None) the cover type.
}}

\item{understoryType_dd}{A data frame containing understory type (0.5 to 5 m) 
values for in the drawdown zone, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE a character value of C (coniferous), D (deciduous), E 
(broadleaf evergreen), M (mixed), or N (None) the cover type.
}}

\item{drawdown}{A data frame indicating presence of drawdown at station, 
with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE an integer value, or character value that is castable to an 
integer, indicating drawdown exists at a site.
}}

\item{horizontalDistance_dd}{A data frame containing the horizontal distance 
to the high water mark where drawdown exists, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE an integer value, or character value that is castable to an 
integer, indicating the horizontal distance to the high water mark when
drawdown exists at a site.
}}

\item{createSyntheticCovers}{A logical value, which specifies whether to 
create synthetic cover values as proportions of drawdown and riparian cover.
This argument should be set to FALSE when the data follows the 2007 NLA 
protocol or do not contain drawdown cover data.  The default value is TRUE.}

\item{fillinDrawdown}{A logical value, which specifies whether to use the
DRAWDOWN parameter to fill in unrecorded cover and HORIZ_DIST_DD values.
The default value is TRUE.}
}
\value{
Either a data frame when metric calculation is successful or a 
character string containing an error message when metric calculation 
is not successful. The data frame contains the following columns:
\itemize{ 
    \item SITE - unique site visit identifier
    \item METRIC - metric name
    \item VALUE - metric value
      }
The output metrics include:
RVFCCANBIG_DD, RVFCCANBIG_RIP, RVFCCANBIG_SYN, RVFCCANSMALL_DD, 
RVFCCANSMALL_RIP, RVFCCANSMALL_SYN, RVFCGNDBARE_DD, RVFCGNDBARE_RIP, 
RVFCGNDBARE_SYN, RVFCGNDINUNDATED_DD, RVFCGNDINUNDATED_RIP, 
RVFCGNDINUNDATED_SYN, 
RVFCGNDNONW_DD, RVFCGNDNONW_RIP, RVFCGNDNONW_SYN, RVFCGNDWOODY_DD, 
RVFCGNDWOODY_RIP, RVFCGNDWOODY_SYN, RVFCUNDNONW_DD, RVFCUNDNONW_RIP, 
RVFCUNDNONW_SYN, RVFCUNDWOODY_DD, RVFCUNDWOODY_RIP, RVFCUNDWOODY_SYN, 
RVFPCANBIG_DD, RVFPCANBIG_RIP, RVFPCANBIG_SYN, RVFPCANBROADLEAF_DD, 
RVFPCANBROADLEAF_RIP, RVFPCANCONIFEROUS_DD, RVFPCANCONIFEROUS_RIP, 
RVFPCANDECIDUOUS_DD, 
RVFPCANDECIDUOUS_RIP, RVFPCANMIXED_DD, RVFPCANMIXED_RIP, RVFPCANNONE_DD, 
RVFPCANNONE_RIP, RVFPCANSMALL_DD, RVFPCANSMALL_RIP, RVFPCANSMALL_SYN, 
RVFPGNDBARE_DD, RVFPGNDBARE_RIP, RVFPGNDBARE_SYN, RVFPGNDINUNDATED_DD, 
RVFPGNDINUNDATED_RIP, RVFPGNDINUNDATED_SYN, RVFPGNDNONW_DD, RVFPGNDNONW_RIP, 
RVFPGNDNONW_SYN, RVFPGNDWOODY_DD, RVFPGNDWOODY_RIP, RVFPGNDWOODY_SYN, 
RVFPUNDBROADLEAF_DD, RVFPUNDBROADLEAF_RIP, RVFPUNDCONIFEROUS_DD, 
RVFPUNDCONIFEROUS_RIP
RVFPUNDDECIDUOUS_DD, RVFPUNDDECIDUOUS_RIP, RVFPUNDMIXED_DD, RVFPUNDMIXED_RIP, 
RVFPUNDNONE_DD, RVFPUNDNONE_RIP, RVFPUNDNONW_DD, RVFPUNDNONW_RIP, 
RVFPUNDNONW_SYN, RVFPUNDWOODY_DD, RVFPUNDWOODY_RIP, RVFPUNDWOODY_SYN, 
RVICANOPY_DD, RVICANOPY_RIP, RVICANOPY_SYN, RVICANUND_DD, 
RVICANUND_RIP, RVICANUND_SYN, RVIGROUND_DD, RVIGROUND_RIP, 
RVIGROUND_SYN, RVIHERBS_DD, RVIHERBS_RIP, RVIHERBS_SYN, 
RVITALLWOOD_DD, RVITALLWOOD_RIP, RVITALLWOOD_SYN, RVITOTALVEG_DD, 
RVITOTALVEG_RIP, RVITOTALVEG_SYN, RVIUNDERSTORY_DD, RVIUNDERSTORY_RIP, 
RVIUNDERSTORY_SYN, RVIWOODY_DD, RVIWOODY_RIP, RVIWOODY_SYN, 
RVNCANBIG_DD, RVNCANBIG_RIP, RVNCANBIG_SYN, RVNCANOPY_DD, 
RVNCANOPY_RIP, RVNCANSMALL_DD, RVNCANSMALL_RIP, RVNCANSMALL_SYN, 
RVNGNDBARE_DD, RVNGNDBARE_RIP, RVNGNDBARE_SYN, RVNGNDINUNDATED_DD, 
RVNGNDINUNDATED_RIP, RVNGNDINUNDATED_SYN, RVNGNDNONW_DD, RVNGNDNONW_RIP, 
RVNGNDNONW_SYN, RVNGNDWOODY_DD, RVNGNDWOODY_RIP, RVNGNDWOODY_SYN, 
RVNUNDERSTORY_DD, RVNUNDERSTORY_RIP, RVNUNDNONW_DD, RVNUNDNONW_RIP, 
RVNUNDNONW_SYN, RVNUNDWOODY_DD, RVNUNDWOODY_RIP, RVNUNDWOODY_SYN, 
RVVCANBIG_DD, RVVCANBIG_RIP, RVVCANBIG_SYN, RVVCANSMALL_DD, 
RVVCANSMALL_RIP, RVVCANSMALL_SYN, RVVGNDBARE_DD, RVVGNDBARE_RIP, 
RVVGNDBARE_SYN, RVVGNDINUNDATED_DD, RVVGNDINUNDATED_RIP, RVVGNDINUNDATED_SYN, 
RVVGNDNONW_DD, RVVGNDNONW_RIP, RVVGNDNONW_SYN, RVVGNDWOODY_DD, 
RVVGNDWOODY_RIP, RVVGNDWOODY_SYN, RVVUNDNONW_DD, RVVUNDNONW_RIP, 
RVVUNDNONW_SYN, RVVUNDWOODY_DD, RVVUNDWOODY_RIP, RVVUNDWOODY_SYN.
 
\emph{NLA_Physical_Habitat_Metric_Descriptions.pdf} in the package
documentation.
}
\description{
This function calculates the riparian zone and 
vegetation portion of the physical habitat metrics for NLA data.  
The function requires a data frame containing validated physical 
habitat data collected using the NLA protocol.
}
\examples{
  head(nlaPhabEx)
  
  bigTrees <- subset(nlaPhabEx,PARAMETER=='C_BIGTREES',select=-PARAMETER)
  bigTrees_dd <- subset(nlaPhabEx,PARAMETER=='C_BIGTREES_DD',select=-PARAMETER)
  smallTrees <- subset(nlaPhabEx,PARAMETER=='C_SMALLTREES',select=-PARAMETER)
  smallTrees_dd <- subset(nlaPhabEx,PARAMETER=='C_SMALLTREES_DD',select=-PARAMETER)
  canopyType <- subset(nlaPhabEx,PARAMETER=='CANOPY',select=-PARAMETER)
  canopyType_dd <- subset(nlaPhabEx,PARAMETER=='CANOPY_DD',select=-PARAMETER)
  grdcvrBare <- subset(nlaPhabEx,PARAMETER=='GC_BARE',select=-PARAMETER)
  grdcvrBare_dd <- subset(nlaPhabEx,PARAMETER=='GC_BARE_DD',select=-PARAMETER)
  grdcvrInund <- subset(nlaPhabEx,PARAMETER=='GC_INUNDATED',select=-PARAMETER)
  grdcvrInund_dd <- subset(nlaPhabEx,PARAMETER=='GC_INUNDATED_DD',select=-PARAMETER)
  grdcvrNw <- subset(nlaPhabEx,PARAMETER=='GC_NONWOODY',select=-PARAMETER)
  grdcvrNw_dd <- subset(nlaPhabEx,PARAMETER=='GC_NONWOODY_DD',select=-PARAMETER)
  grdcvrW <- subset(nlaPhabEx,PARAMETER=='GC_WOODY',select=-PARAMETER)
  grdcvrW_dd <- subset(nlaPhabEx,PARAMETER=='GC_WOODY_DD',select=-PARAMETER)
  undNonW <- subset(nlaPhabEx,PARAMETER=='U_NONWOODY',select=-PARAMETER)
  undNonW_dd <- subset(nlaPhabEx,PARAMETER=='U_NONWOODY_DD',select=-PARAMETER)
  undW <- subset(nlaPhabEx,PARAMETER=='U_WOODY',select=-PARAMETER)
  undW_dd <- subset(nlaPhabEx,PARAMETER=='U_WOODY_DD',select=-PARAMETER)
  undType <- subset(nlaPhabEx,PARAMETER=='UNDERSTORY',select=-PARAMETER)
  undType_dd <- subset(nlaPhabEx,PARAMETER=='UNDERSTORY_DD',select=-PARAMETER)
  drawdown <- subset(nlaPhabEx,PARAMETER=='DRAWDOWN',select=-PARAMETER)
  horizontalDistance_dd <- subset(nlaPhabEx,PARAMETER=='HORIZ_DIST_DD',select=-PARAMETER)
  
  # Use defaults for fillinDrawdown and createSyntheticCovers
  exRipVeg <- nlaRiparianVegetation(bigTrees, bigTrees_dd, smallTrees, smallTrees_dd,
  canopyType, canopyType_dd, grdcvrBare, grdcvrBare_dd, grdcvrInund, grdcvrInund_dd,
  grdcvrNw, grdcvrNw_dd, grdcvrW, grdcvrW_dd, undNonW, undNonW_dd, undW, undW_dd,
  undType, undType_dd, drawdown, horizontalDistance_dd)
  
  head(exRipVeg)
 
}
\author{
Curt Seeliger \email{Seeliger.Curt@epa.gov}\cr
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
