% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLAphab_indicators.r
\name{nlaRipDistIndicator}
\alias{nlaRipDistIndicator}
\title{Calculate NLA Lakeshore Anthropogenic Disturbance Indicator}
\usage{
nlaRipDistIndicator(x, sampID, hiiAg, hiiNonAg, hifpAnyCirca)
}
\arguments{
\item{x}{Input data frame containing other variables as specified in
arguments}

\item{sampID}{Vector of variables necessary to identify a unique sample 
in \emph{x}.}

\item{hiiAg}{NLA physical habitat index of agricultural influences
based on weighted mean human influence metrics for crops, pasture, 
and orchards. Calculated by the function \code{metsHumanImpact()}.}

\item{hiiNonAg}{NLA physical habitat index of non-agricultural influences
based on weighted mean human influence metrics for buildings, commercial,
docks, landfill, lawn, parks, powerlines, roads, and walls. Calculated
by the function \code{metsHumanImpact()}.}

\item{hifpAnyCirca}{Fractional cover of any human influence found within the 
plot area at any station. Calculated by the function \code{metsHumanImpact()}}
}
\value{
A data frame containing:
\itemize{
\item{sampID}{The variables in the argument \emph{sampID}}

\item{RDis_IX}{Riparian anthropogenic disturbance index}

\item{RDIS_COND}{Riparian anthropogenic disturbance condition class
 (Low/Medium/High/Not Assessed)} 
}
}
\description{
Using metric values as inputs, calculate 
indicator score for RDis_IX, the indicator of lakeshore anthropogenic disturbance. 
Use thresholds based on reference distributions across all ecoregions.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
\keyword{survey}
