% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLAphab_indicators.r
\name{nlaRipVegCompIndicator}
\alias{nlaRipVegCompIndicator}
\title{Calculate NLA Riparian Vegetation Complexity Indicator}
\usage{
nlaRipVegCompIndicator(x, sampID, lat, lon, lake_origin, area, elev, ecoreg,
  rviWoody, rvfcGndInundated, rvfcUndWoody, rvfcGndWoody, rvfpCanBig,
  ssfcBedrock, ssfcBoulders, hipwWalls)
}
\arguments{
\item{x}{Input data frame containing other variables as specified in
arguments}

\item{sampID}{Vector of variables necessary to identify a unique sample 
in \emph{x}.}

\item{lat}{Latitude of lake in decimal degrees (NAD 83 datum)}

\item{lon}{Longitude of lake in decimal degrees (NAD 83 datum)}

\item{lake_origin}{Lake origin, with valid values of 'NATURAL' 
or 'MAN_MADE'}

\item{area}{Area of lake or reservoir, in square kilometers}

\item{elev}{Lake surface elevation (meters above sea level)}

\item{ecoreg}{Lake ecoregion, based on aggregated Omernick ecoregions,
with valid values of CPL, NAP, NPL, SAP, SPL, TPL, UMW, WMT, XER.}

\item{rviWoody}{Index of total woody vegetation cover, calculated
by function \code{metsRiparianVegetationNLA()}.}

\item{rvfcGndInundated}{Fraction of ground cover inundated, 
calculated by function \code{metsRiparianVegetationNLA()}.}

\item{rvfcUndWoody}{Fraction of understory cover as woody vegetation,
calculated by function \code{metsRiparianVegetationNLA()}}

\item{rvfcGndWoody}{Fraction of ground cover as woody vegetation,
calculated by function \code{metsRiparianVegetationNLA()}}

\item{rvfpCanBig}{Fraction of canopy as trees with dbh > 30 cm,
calculated by function \code{metsRiparianVegetationNLA()}}

\item{ssfcBedrock}{Fractional shoreline cover of bedrock, calculated
by function \code{metsShorelineSubstrate()}.}

\item{ssfcBoulders}{Fractional shoreline cover of boulders, calculated
by function \code{metsShorelineSubstrate()}.}

\item{hipwWalls}{Human influence weighted presence of walls, calculated 
by function \code{metsHumanImpact()}.}
}
\value{
A data frame containing:
\itemize{
\item{sampID}{The variables in the argument \emph{sampID}}

\item{RVegQ}{Observed riparian vegetation complexity index}

\item{RVegQc3x15}{Expected riparian vegetation complexity index}

\item{RVegQc3OE}{Riparian vegetation complexity indicator 
value, O/E score}

\item{RVEG_COND}{Riparian vegetation complexity indicator 
condition class (Good/Fair/Poor/Not Assessed)} 
}
}
\description{
Using metric values and various predictors as inputs, calculate 
indicator score for RVegQc3OE, the observed over expected value for
riparian vegetation complexity. Models to calculate expected values depend
on ecoregion. Use thresholds based on reference distributions by ecoregion to 
assign condition.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
\keyword{survey}
