% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrsaChanconst}
\alias{nrsaChanconst}
\title{Example data for NRSA channel constraint characteristics}
\format{
a data frame with 25 rows and 4 variables
\describe{
 \item{SITE}{Site visit identifier}
 \item{SAMPLE_TYPE}{Sample type, corresponding to field form used, value
 CONSTRAINT}
 \item{PARAMETER}{Parameter name for field measurment}
 \item{VALUE}{Value of parameter}
}
}
\source{
aquamet package
}
\usage{
nrsaChanconst
}
\description{
Example data for NRSA channel constraint characteristics
}
\keyword{datasets}
