% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrsaChannel}
\alias{nrsaChannel}
\title{Example data for NRSA channel measurements}
\format{
a data frame with 1865 rows and 5 variables
\describe{
 \item{SITE}{Site visit identifier}
 \item{TRANSECT}{Transect value, range A-K}
 \item{SAMPLE_TYPE}{Sample type, corresponding to field form used, values
 BANKW, FISHCOVW, LWDW, BANKB, CONSTB, FISHCOVB, LITTORALB, LWDB}
 \item{PARAMETER}{Parameter name for field measurment}
 \item{VALUE}{Value of parameter}
}
}
\source{
aquamet package
}
\usage{
nrsaChannel
}
\description{
Example data for NRSA channel measurements
}
\keyword{datasets}
