% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrsaChannelMorphology.r
\name{nrsaChannelMorphology}
\alias{nrsaChannelMorphology}
\title{Calculate NRSA Channel Morphology Metrics}
\usage{
nrsaChannelMorphology(
  bBankHeight = NULL,
  bBankWidth = NULL,
  bDepth = NULL,
  bIncisedHeight = NULL,
  bWettedWidth = NULL,
  wBankHeight = NULL,
  wBankWidth = NULL,
  wDepth = NULL,
  wIncisedHeight = NULL,
  wWettedWidth = NULL,
  isUnitTest = FALSE
)
}
\arguments{
\item{bBankHeight}{A data frame containing bank height at each transect for
  boatable reaches, with the following columns:
 \itemize{
 \item SITE        integer or character specifying the site visit
 \item TRANSECT    character value specifying the transect
                        for which the value was recorded.
 \item VALUE       numeric or character values
 \item UNITS       character value specifying the units that
                             the value is recorded in, expected to be
                             either CM, M or FT.
}}

\item{bBankWidth}{A data frame containing bank width at each transect for
boatable reaches, with the following columns:
\itemize{
 \item SITE        integer or character specifying the site visit
 \item TRANSECT    character value specifying the transect
            for which the value was recorded.
 \item VALUE       numeric or character values
 \item UNITS       character value specifying the units that
                      the value is recorded in, expected to be
                      either CM, M or FT.
}}

\item{bDepth}{A data frame containing thalweg depth at each transect for
boatable reaches, with the following columns:
\itemize{
     \item SITE        integer or character specifying the site visit
     \item TRANSECT    character value specifying the transect
                      for which the value was recorded.
     \item STATION     numeric value specifying the station between
                     the transects at which the depth was recorded.
     \item VALUE       numeric or character values
     \item UNITS       character value specifying the units that
                      the value is recorded in, expected to be
                      either CM, M or FT.
}}

\item{bIncisedHeight}{A data frame containing incised height at each transect for
boatable reaches, with the following columns:
\itemize{
     \item SITE        integer or character specifying the site visit
     \item TRANSECT    character value specifying the transect
                      for which the value was recorded.
     \item VALUE       numeric or character values
     \item UNITS       character value specifying the units that
                      the value is recorded in, expected to be
                      either CM, M or FT.
}}

\item{bWettedWidth}{A data frame containing wetted width at each transect for
boatable reaches, with the following columns:
\itemize{
 \item SITE integer or character specifying the site visit
 \item TRANSECT character value specifying the transect
               for which the value was recorded.
 \item VALUE numeric or character values
 \item UNITS character value specifying the units that
             the value is recorded in, expected to be
             either CM, M or FT.
}}

\item{wBankHeight}{A data frame containing bank height at each transect for
 wadeable reaches, with the following columns:
 \itemize{
          \item SITE integer or character specifying the site visit
          \item TRANSECT character value specifying the transect
                         for which the value was recorded.
          \item VALUE numeric or character values
          \item UNITS character value specifying the units that
                      the value is recorded in, expected to be
                      either CM, M or FT.
}}

\item{wBankWidth}{A data frame containing bank width at each transect for
wadeable reaches, with the following columns:
\itemize{
     \item SITE integer or character specifying the site visit
     \item TRANSECT character value specifying the transect
                      for which the value was recorded.
     \item VALUE numeric or character values
     \item UNITS character value specifying the units that
                 the value is recorded in, expected to be
                 either CM, M or FT.
}}

\item{wDepth}{A data frame containing thalweg depth at each transect for
wadeable reaches, with the following columns:
\itemize{
   \item SITE        integer or character specifying the site visit
   \item TRANSECT    character value specifying the transect
                    for which the value was recorded.
   \item STATION     numeric value specifying the station between
                    the transects at which the depth was recorded.
   \item VALUE       numeric or character values
   \item UNITS       character value specifying the units that
                    the value is recorded in, expected to be
                    either CM, M or FT.
}}

\item{wIncisedHeight}{A data frame containing incised height at each 
transect for wadeable reaches, with the following columns:
\itemize{
     \item SITE integer or character specifying the site visit
     \item TRANSECT character value specifying the transect
                     for which the value was recorded.
     \item VALUE numeric or character values
     \item UNITS character value specifying the units that
                 the value is recorded in, expected to be
                 either CM, M or FT.
}}

\item{wWettedWidth}{A data frame containing wetted width at each transect 
for wadeable reaches, with the following columns:
\itemize{
         \item SITE integer or character specifying the site visit
         \item TRANSECT character value specifying the transect
                        for which the value was recorded.
         \item STATION numeric value specifying the station between
                    the transects at which the depth was recorded.
         \item VALUE numeric or character values
         \item UNITS character value specifying the units that
                     the value is recorded in, expected to be
                     either CM, M or FT.

}}

\item{isUnitTest}{Logical argument to determine whether errors should be ignored.
Should only be used for running a unit test. Default value is FALSE.}
}
\value{
Either a data frame when metric calculation is successful or a 
character string containing an error message when metric calculation is 
not successful.  The data frame contains the following columns:
\itemize{
    \item SITE - universal ID value
    \item METRIC - metric name
    \item VALUE - metric value
}

Wadeable stream metrics calculated include: bfwd_rat, n_bfrat, n_bh, 
n_bw, n_d, n_incis, n_w, n_wd,n_wdr, sdbkf_h, sdbkf_w, sddepth, 
sdinc_h, sdwd_rat, sdwidth, sdwxd,xbkf_h, xbkf_w, xdepth, 
xinc_h, xwd_rat, xwidth, xwxd

Boatable river metrics calculated include:  bfwd_rat, n_bfrat, n_bh, 
n_bw, n_d, n_incis,  n_wd,n_wdr,sdbkf_h, sdbkf_w, sddepth, sdinc_h, 
sdwd_rat, sdwidth, sdwxd,xbkf_h, xbkf_w, xdepth, xinc_h, 
xwd_rat, xwidth, xwxd
Descriptions for all metrics are included in 
\emph{NRSA_Physical_Habitat_Metric_Descriptions.pdf} in the package
documentation.
}
\description{
This function calculates the channel morphology 
portion of the physical habitat metrics for National Rivers 
and Streams Assessment (NRSA) data.  The function requires data 
frames containing the bank geometry, thalweg, and stream 
verification form data files.
}
\examples{
head(bankgeomEx)
head(thalwegEx)

# Must subset example dataset to create inputs, keeping only SITE, TRANSECT,
  #  VALUE, and UNITS, then make sure VALUE is numeric.
# bBankHeight 
bBH <- subset(bankgeomEx,PARAMETER=='BANKHT' & SAMPLE_TYPE=='PHAB_CHANBFRONT',
        select=c(SITE,TRANSECT,VALUE,UNITS))
        bBH$VALUE <- as.numeric(bBH$VALUE)
# bBankWidth
bBW <- subset(bankgeomEx,PARAMETER=='BANKWID' & SAMPLE_TYPE=='PHAB_CHANBFRONT',
        select=c(SITE,TRANSECT,VALUE,UNITS))
        bBW$VALUE <- as.numeric(bBW$VALUE)
# bDepth
bD <- subset(thalwegEx,(PARAMETER=='DEP_SONR'|PARAMETER=='DEP_POLE') & 
        SAMPLE_TYPE=='PHAB_THAL',select=c(SITE,STATION,TRANSECT,VALUE,UNITS))
        bD$VALUE <- as.numeric(bD$VALUE)
# bIncisedHeight
bInc <- subset(bankgeomEx,PARAMETER=='INCISED' & SAMPLE_TYPE=='PHAB_CHANBFRONT',
        select=c(SITE,TRANSECT,VALUE,UNITS))
        bInc$VALUE <- as.numeric(bInc$VALUE)
# bWettedWidth
bWW <- subset(bankgeomEx,PARAMETER=='WETWID' & SAMPLE_TYPE=='PHAB_CHANBFRONT',
        select=c(SITE,TRANSECT,VALUE,UNITS))
        bWW$VALUE <- as.numeric(bWW$VALUE)
# wBankHeight
wBH <- subset(bankgeomEx,PARAMETER=='BANKHGT' & SAMPLE_TYPE=='PHAB_CHANW',
        select=c(SITE,TRANSECT,VALUE,UNITS))
        wBH$VALUE <- as.numeric(wBH$VALUE)
# wBankWidth
wBW <- subset(bankgeomEx,PARAMETER=='BANKWID' & SAMPLE_TYPE=='PHAB_CHANW',
        select=c(SITE,TRANSECT,VALUE,UNITS))
        wBW$VALUE <- as.numeric(wBW$VALUE)
# wDepth
wD <- subset(thalwegEx,PARAMETER=='DEPTH' & SAMPLE_TYPE=='PHAB_THALW',
        select=c(SITE,STATION,TRANSECT,VALUE,UNITS))
        wD$VALUE <- as.numeric(wD$VALUE)
# wIncisedHeight
wInc <- subset(bankgeomEx,PARAMETER=='INCISHGT' & SAMPLE_TYPE=='PHAB_CHANW',
        select=c(SITE,TRANSECT,VALUE,UNITS))
        wInc$VALUE <- as.numeric(wInc$VALUE)
# wWettedWidth
wWW <- subset(thalwegEx,PARAMETER=='WETWIDTH', 
        select=c(SITE,STATION,TRANSECT,VALUE,UNITS))
        wWW$VALUE <- as.numeric(wWW$VALUE)

chanmorphOut <- nrsaChannelMorphology(bBankHeight=bBH, bBankWidth=bBW,
bDepth=bD, bIncisedHeight=bInc, bWettedWidth=bWW, wBankHeight=wBH,
wBankWidth=wBW, wDepth=wD, wIncisedHeight=wInc, wWettedWidth=wWW)

head(chanmorphOut)
}
\author{
Curt Seeliger \email{Seeliger.Curt@epa.gov}\cr
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
