% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrsaChanrip}
\alias{nrsaChanrip}
\title{Example data for NRSA visual riparian estimates}
\format{
a data frame with 2200 rows and 6 variables
\describe{
 \item{SITE}{Site visit identifier}
 \item{TRANSECT}{Transect value, range A-K}
 \item{BANK}{Bank where measurement taken, values, LF, RT}
 \item{SAMPLE_TYPE}{Sample type, corresponding to field form used, values
 PHAB_CHANW and PHAB_CHANB}
 \item{PARAMETER}{Parameter name for field measurment}
 \item{VALUE}{Value of parameter}
}
}
\source{
ALL_THE_NRSA database
}
\usage{
nrsaChanrip
}
\description{
Example data for NRSA visual riparian estimates
}
\keyword{datasets}
