% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrsaChanxsec}
\alias{nrsaChanxsec}
\title{Example data for NRSA channel cross-section measurements for wadeable streams}
\format{
a data frame with 560 rows and 6 variables
\describe{
 \item{SITE}{Site visit identifier}
 \item{TRANSECT}{Transect value, range A-K}
 \item{BANK}{Bank where measurement taken, values, LF, RT, NONE}
 \item{SAMPLE_TYPE}{Sample type, corresponding to field form used, values
 PHAB_CHANW and PHAB_THALW}
 \item{PARAMETER}{Parameter name for field measurment}
 \item{VALUE}{Value of parameter}
}
}
\source{
aquamet package
}
\usage{
nrsaChanxsec
}
\description{
Example data for NRSA channel cross-section measurements for wadeable streams
}
\keyword{datasets}
