% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrsaInvleg}
\alias{nrsaInvleg}
\title{Example data for NRSA legacy riparian vegetation}
\format{
a data frame with 264 rows and 4 variables
\describe{
 \item{SITE}{Site visit identifier}
 \item{SAMPLE_TYPE}{Sample type: RIPLEG}
 \item{TRANSECT}{Transect value, range A-K}
 \item{PARAMETER}{Parameter name for field measurment}
 \item{VALUE}{Value of parameter}
}
}
\source{
aquamet package
}
\usage{
nrsaInvleg
}
\description{
Example data for NRSA legacy riparian vegetation
}
\keyword{datasets}
