% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrsaLargeWoody.r
\name{nrsaLargeWoody}
\alias{nrsaLargeWoody}
\title{Calculate NRSA Large Woody Debris Metrics}
\usage{
nrsaLargeWoody(bCounts = NULL, bClassInfo = data.frame(diam = rep(c("SMALL",
  "MEDIUM", "LARGE", "EXTRALARGE"), each = 3, times = 2), len = rep(c("SMALL",
  "MEDIUM", "LARGE"), times = 4 * 2), loc = rep(c("DRY", "WET"), each = 12),
  stringsAsFactors = FALSE) \%>\% mutate(CLASS = sprintf("\%s\%sD\%sL",
  substr(loc, 1, 1), ifelse(substr(diam, 1, 1) == "E", "X", substr(diam, 1, 1)),
  substr(len, 1, 1))), wCounts = NULL, wClassInfo = data.frame(diam =
  rep(c("SMALL", "MEDIUM", "LARGE", "EXTRALARGE"), each = 3, times = 2), len =
  rep(c("SMALL", "MEDIUM", "LARGE"), times = 4 * 2), loc = rep(c("DRY", "WET"),
  each = 12), stringsAsFactors = FALSE) \%>\% mutate(CLASS =
  sprintf("\%s\%sD\%sL", substr(loc, 1, 1), ifelse(substr(diam, 1, 1) == "E",
  "X", substr(diam, 1, 1)), substr(len, 1, 1))), reachlength = NULL,
  meanBankfullWidth = NULL)
}
\arguments{
\item{bCounts}{A data frame containing large woody debris class counts 
at each transect for boatable reaches , with the following columns:
\itemize{
   \item SITE integer or character specifying the site visit
   \item TRANSECT character value specifying the transect
                  for which the value was recorded.
   \item CLASS character value specifying the class name
               (one of 4 diameter classes, 3 length classes,
               and 2 location classes).  These class names are
               expected to be defined in the bClassInfo
               argument
   \item VALUE character values specifying the influence 
               class
}}

\item{bClassInfo}{A data frame relating the class names used in 
bCounts::CLASS to each combination of size and location classes.  
Expected to have the following columns and values in 24 rows:
\itemize{
     \item diam character value specifying diameter class: SMALL
                MEDIUM, LARGE, EXTRALARGE, as defined in the NRSA
                field protocol manual
     \item len character value specifying length class: SMALL
               MEDIUM, LARGE, as defined in the NRSA field protocol
               manual
     \item loc character value specifying location class: DRY, WET
               as defined in the NRSA field protocol manual
     \item CLASS character value associated with each combination of
                 diameter, length and channel location, as it occurs
                 in the bCounts argument.
}
Default values are for EPA NARS processing.}

\item{wCounts}{A data frame containing large woody debris class counts 
at each transect for wadeable reaches , with the following columns:
\itemize{
   \item SITE integer or character specifying the site visit
   \item TRANSECT character value specifying the transect
                  for which the value was recorded.
   \item CLASS character value specifying the class name
               (one of 4 diameter classes, 3 length classes,
               and 2 location classes).  These class names are
               expected to be defined in the wClassInfo
               argument
   \item VALUE character values specifying the influence 
               class
}}

\item{wClassInfo}{A data frame relating the class names used in 
wCounts::CLASS to each combination of size and location classes.  
Expected to have the following columns and values in 24 rows:
\itemize{
   \item diam character value specifying diameter class: SMALL
              MEDIUM, LARGE, EXTRALARGE, as defined in the NRSA
              field protocol manual
   \item len character value specifying length class: SMALL
             MEDIUM, LARGE, as defined in the NRSA field protocol
             manual
   \item loc character value specifying location class: DRY, WET
             as defined in the NRSA field protocol manual
   \item CLASS character value associated with each combination of
               diameter, length and channel location, as it occurs
               in the wCounts argument.
}
Default value is for EPA NARS processing.}

\item{reachlength}{A data frame containing reach lengths in meters 
for each site. Expected to contain the following columns:
\itemize{
   \item SITE integer or character specifying the site visit
   \item VALUE numeric values
}}

\item{meanBankfullWidth}{A data frame containing mean bank-full 
widths in meters for each site.  Expected to contain the following 
columns:
\itemize{
 \item SITE integer or character specifying the site visit
 \item VALUE numeric values

}}
}
\value{
Either a data frame when metric calculation is successful 
or a character string containing an error message when metric 
calculation is not successful.  The data frame contains the following 
columns:
\itemize{
   \item SITE - universal ID value
   \item METRIC - metric name
   \item VALUE - metric value
}
Metrics calculated for wadeable sites include: 
c1d, c1dm100, c1t, c1tm100, c1w, c1w_msq, c1wm100, c2d, c2dm100, c2t,
  c2tm100, c2w, c2w_msq, c2wm100, c3d, c3dm100, c3t, c3tm100, c3w, c3w_msq,
  c3wm100, c4d, c4dm100, c4t, c4tm100, c4w, c4w_msq, c4wm100, c5d, c5dm100,
  c5t, c5tm100, c5w, c5w_msq, c5wm100, lgdiatot, lgdrydia, lgdrylen, lglentot,
  lgwetdia, lgwetlen, lwddv33, lwddvcal, lwdtv33, lwdtvcal, lwdwv33, lwdwvcal,
  mddiatot, mddrydia, mddrylen, mdlentot, mdwetdia, mdwetlen, nc, ns,
  rchdldll, rchdldml, rchdldsl, rchdmdll, rchdmdml, rchdmdsl, rchdryt,
  rchdsdll, rchdsdml, rchdsdsl, rchdxdll, rchdxdml, rchdxdsl, rchtldll,
  rchtldml, rchtldsl, rchtmdll, rchtmdml, rchtmdsl, rchtsdll, rchtsdml,
  rchtsdsl, rchtxdll, rchtxdml, rchtxdsl, rchwdt, rchwett, rchwldll, rchwldml,
  rchwldsl, rchwmdll, rchwmdml, rchwmdsl, rchwsdll, rchwsdml, rchwsdsl,
  rchwxdll, rchwxdml, rchwxdsl, shdrylen, shlentot, shwetlen, smdiatot,
  smdrydia, smwetdia, v1d, v1dm100, v1t, v1tm100, v1w, v1w_msq, v1wm100, v2d,
  v2dm100, v2t, v2tm100, v2w, v2w_msq, v2wm100, v3d, v3dm100, v3t, v3tm100,
  v3w, v3w_msq, v3wm100, v4d, v4dm100, v4t, v4tm100, v4w, v4w_msq, v4wm100,
  v5d, v5dm100, v5t, v5tm100, v5w, v5w_msq, v5wm100, xldiatot, xldrydia,
  xlwetdia
  
Metrics calculated for boatable sites include:
  c1d, c1dm100, c1t, c1tm100, c1w, c1w_msq, c1wm100, c2d, c2dm100, c2t,
  c2tm100, c2w, c2w_msq, c2wm100, c3d, c3dm100, c3t, c3tm100, c3w, c3w_msq,
  c3wm100, c4d, c4dm100, c4t, c4tm100, c4w, c4w_msq, c4wm100, c5d, c5dm100,
  c5t, c5tm100, c5w, c5w_msq, c5wm100, lgdiatot, lgdrydia, lgdrylen, lglentot,
  lgwetdia, lgwetlen, lwddv33, lwddvcal, lwdtv33, lwdtvcal, lwdwv33, lwdwvcal,
  mddiatot, mddrydia, mddrylen, mdlentot, mdwetdia, mdwetlen, nc, ns,
  rchdldll, rchdldml, rchdldsl, rchdmdll, rchdmdml, rchdmdsl, rchdryt,
  rchdsdll, rchdsdml, rchdsdsl, rchdxdll, rchdxdml, rchdxdsl, rchtldll,
  rchtldml, rchtldsl, rchtmdll, rchtmdml, rchtmdsl, rchtsdll, rchtsdml,
  rchtsdsl, rchtxdll, rchtxdml, rchtxdsl, rchwdt, rchwett, rchwldll, rchwldml,
  rchwldsl, rchwmdll, rchwmdml, rchwmdsl, rchwsdll, rchwsdml, rchwsdsl,
  rchwxdll, rchwxdml, rchwxdsl, shdrylen, shlentot, shwetlen, smdiatot,
  smdrydia, smwetdia, v1d, v1dm100, v1t, v1tm100, v1w, v1w_msq, v1wm100, v2d,
  v2dm100, v2t, v2tm100, v2w, v2w_msq, v2wm100, v3d, v3dm100, v3t, v3tm100,
  v3w, v3w_msq, v3wm100, v4d, v4dm100, v4t, v4tm100, v4w, v4w_msq, v4wm100,
  v5d, v5dm100, v5t, v5tm100, v5w, v5w_msq, v5wm100, xldiatot, xldrydia,
  xlwetdia 
      
Descriptions for all metrics are included in 
\emph{NRSA_Physical_Habitat_Metric_Descriptions.pdf} in the package
documentation.
}
\description{
This function calculates the large woody debris 
portion of the physical habitat metrics for National Rivers 
and Streams Assessment (NRSA) data.  The function requires data 
frames containing the thalweg, channel geometry, bank 
geometry, large woody debris, and stream verification form data files.
}
\examples{
# See Bed Stability vignette for example of calculating large woody debris metrics.
  
}
\author{
Curt Seeliger \email{Seeliger.Curt@epa.gov}\cr
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
