% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrsaLittoral}
\alias{nrsaLittoral}
\title{Example data for NRSA channel depth measurements in non-wadeable sites}
\format{
a data frame with 132 rows and 5 variables
\describe{
 \item{SITE}{Site visit identifier}
 \item{SAMPLE_TYPE}{Sample type: CHANDEPTHB}
 \item{TRANSECT}{Transect value, range A-K}
 \item{LINE}{Line number from form}
 \item{PARAMETER}{Parameter name for field measurment}
 \item{VALUE}{Value of parameter}
}
}
\source{
aquamet package
}
\usage{
nrsaLittoral
}
\description{
Example data for NRSA channel depth measurements in non-wadeable sites
}
\keyword{datasets}
