% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrsaLittoralDepth.r
\name{nrsaLittoralDepth}
\alias{nrsaLittoralDepth}
\title{Calculate NRSA Littoral Depth Metrics}
\usage{
nrsaLittoralDepth(
  bLittoralDepth = NULL,
  isUnitTest = FALSE,
  argSavePath = NULL
)
}
\arguments{
\item{bLittoralDepth}{A data frame containing littoral depth 
data in meters at each transect for all reaches, with the 
following columns:
\itemize{
 \item SITE integer or character specifying the site visit
 \item VALUE numeric or character values
}
Note that possible values for variables in the input data 
frame are provided in the document named "NRSA Documentation.pdf" 
included in the help directory for the package.}

\item{isUnitTest}{Logical argument to determine whether errors should be ignored.
Should only be used for running a unit test. Default value is FALSE.}

\item{argSavePath}{character string specifying the path to which data 
arguments are saved as csv files, or NULL if those files are not to be 
written.}
}
\value{
Either a data frame when metric calculation is successful 
or a character string containing an error message when metric 
calculation is not successful.  The data frame contains the following 
columns:
\itemize{
   \item SITE - universal ID value
   \item METRIC - metric name
   \item VALUE - metric value
}
Metrics calculated (only for boatable sites) include: xlit, mxlit, 
mnlit, vlit. 

Descriptions for all NRSA metrics can be found at:
\href{https://github.com/USEPA/aquamet/blob/master/inst/NRSA_physical_habitat_metrics_descriptions.pdf}{NLA_Physical_Habitat_Metric_Descriptions.pdf}.
}
\description{
This function calculates the littoral depth 
portion of the physical habitat metrics for National Rivers 
and Streams Assessment (NRSA) data.  The function requires a 
data frame containing the channel depth data file.
}
\examples{
head(chandepthEx) 

# Subset channel depth example dataset to create input, keeping only SITE
  #  and VALUE. This only applies to boatable sites.
bLitDep <- subset(chandepthEx, SAMPLE_TYPE=='PHAB_CHANBFRONT' & 
  PARAMETER \%in\% c('POLE','SONAR'), select = c(SITE,VALUE))
  # Ensure VALUE is numeric.
  bLitDep$VALUE <- as.numeric(bLitDep$VALUE)

litDepOut <- nrsaLittoralDepth(bLitDep)
print(litDepOut)

}
\author{
Curt Seeliger \email{Seeliger.Curt@epa.gov}\cr
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
