% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRSAphab_indicators.r
\name{nrsaRelBedStabilityIndicator}
\alias{nrsaRelBedStabilityIndicator}
\title{Assign NRSA Relative Bed Stability indicator condition class}
\usage{
nrsaRelBedStabilityIndicator(x, sampID = "UID", ecoreg, protocol, lrbs, lat,
  lon, area, elev, slope, xwidth)
}
\arguments{
\item{x}{A data frame containing, at minimum, the variables 
specified in the arguments for \emph{sampID}, ecoreg, realm, lrbs, and 
any variables necessary to calculate indicator condition classes 
for ecoregions in x. The necessary variables by realm and ecoreg 
are (none indicates no additional variables are necessary):
\itemize{
\item CPL, BOATABLE: none
\item CPL, WADEABLE: xwidth, slope
\item NAP, BOATABLE: none
\item NAP, WADEABLE: area
\item NPL, BOATABLE: area, slope
\item NPL, WADEABLE: elev, xwidth, slope
\item SAP, BOATABLE: none
\item SAP, WADEABLE: area
\item SPL, BOATABLE: area
\item SPL, WADEABLE: lat, area, slope
\item TPL, BOATABLE: area
\item TPL, WADEABLE: lat, lon, slope
\item UMW, BOATABLE: lat
\item UMW, WADEABLE: slope
\item WMT, BOATABLE: none
\item WMT, WADEABLE: xwidth, slope
\item XER, BOATABLE: none
\item XER, WADEABLE: xwidth
}}

\item{sampID}{A character vector containing the names of all 
variables in \emph{x} that specify a unique sample. If not specified, 
the default is \emph{UID}.}

\item{ecoreg}{A string with the name of the aggregated ecoregion variable. 
Valid values that correspond to regions used in NRSA are
CPL, NAP, NPL, SAP, SPL, TPL, UMW, WMTNS, and XER.}

\item{protocol}{A string with the name of the variable indicating
physical habitat protocol used. The expected values of this variable are 
'BOATABLE' and 'WADEABLE'.}

\item{lrbs}{A string with the name of the variable for log10(relative bed
stability). Typically, the version used from NRSA physical habitat metrics
is LRBS_G08.}

\item{lat}{A string with the name of the variable for latitude, assumed to 
be in decimal degrees using NAD83.}

\item{lon}{A string with the name of the variable for longitude, assumed to
be in decimal degrees using NAD83.}

\item{area}{A string with the name of the variable for watershed area in km2}

\item{elev}{A string with the name of the variable for elevation in meters. In NRSA, 
the variable used is ELEV_PT.}

\item{slope}{A string with the name of the variable for % stream slope. For 
NRSA, the variable used is XSLOPE.}

\item{xwidth}{A string with the name of the variable for mean wetted width
in meters. For NRSA, the variable used is XWIDTH.}
}
\value{
A data frame containing the variables in \emph{sampID} and BEDSED_COND, 
the condition class for relative bed stability.
}
\description{
Using models based on NRSA 2008-9 and 2013-14 reference sites, 
this function determines the condition class by aggregated ecoregion and
sampling protocol used. The model used to determine the expected Log10(RBS)
(LRBS) depends on the ecoregion and protocol.
}
\references{
Add Phil Kaufmann's technical report here
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
\keyword{survey}
