% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrsaResidualPools.r
\name{nrsaResidualPools}
\alias{nrsaResidualPools}
\title{Calculate NRSA Residual Pools Metrics}
\usage{
nrsaResidualPools(bDepth = NULL, wDepth = NULL, siteSlopes = NULL,
  transectSpacing = NULL, writeIntermediateFiles = FALSE,
  oldeMethods = FALSE)
}
\arguments{
\item{bDepth}{A data frame containing thalweg depths for boatable 
reaches, with the following columns:
\itemize{
     \item SITE integer or character specifying the site visit
     \item TRANSECT character value of single letters 'A' through 'K'
           where 'A' is at the upstream end and 'K' is at the 
           downstream end of the reach.
     \item STATION integer value ranging from 0 to N indicating 
                   location between transects.  The value of 0 occurs
                   at the transect location, followed by 1, and runs
                   to some value N which is located just before the
                   next transect, at which STATION then changes back 
                   to 0.
     \item VALUE numeric value of individual depths
     \item UNITS character value indicating units of depth values;
                 is expected to be 'FT' or 'M'.
}}

\item{wDepth}{A data frame containing thalweg depths for wadeable 
reaches, with the following columns:
 \itemize{
       \item SITE integer or character specifying the site visit
       \item TRANSECT character value of single letters 'A' through 'K'
                       where 'A' is at the downstream end and 'K' is at the 
                       upstream end of the reach.
       \item STATION integer value ranging from 0 to 9 or 14, depending 
                     on stream width, indicating location between 
                     transects.  The value of 0 occurs at the transect 
                     location, followed by 1, and runs to either 9 or 14 
                     which is located just before the next transect, at 
                     which STATION then changes back to 0.
       \item VALUE numeric value of individual depths, always in 
                   centimeters.
}}

\item{siteSlopes}{A data frame containing the mean slope for all reaches, 
with the following columns:
\itemize{
     \item SITE integer or character specifying the site visit
     \item VALUE numeric value of individual slopes, always in 
                 percent.
}}

\item{transectSpacing}{A data frame containing the transect spacing 
for the sites, with the following columns: 
\itemize{
   \item SITE integer or character specifying the site visit
   \item TRANSECT character value of single letters 'A' through 'K'.
   \item VALUE numeric value of distance between transects, always 
               in meters.
}}

\item{writeIntermediateFiles}{logical value; intermediate results will 
be written to the results folder if TRUE.}

\item{oldeMethods}{logical value; if TRUE, uses deprecated SAS method 
of starting initial pool dimensions in metsResidualPools.dimensions(), 
otherwise uses current method.  Set to TRUE only for unit testing.  
IT IS PROBABLY BEST TO NOT INCLUDE THIS IN THE USER DOCUMENTATION.}
}
\value{
Either a data frame when metric calculation is successful 
or a character string containing an error message when metric 
calculation is not successful.  The data frame contains the following 
columns:
\itemize{
   \item SITE - universal ID value
   \item METRIC - metric name
   \item VALUE - metric value
} 
Pool characteristic metrics: poolID, poolar, poolen, mindep, xdep, rpvdep,
 rpmxdep, meddep, dep25, dep75 
 
Site summary metrics: rpxlen, rpvlen, rpmxlen, totplen, rpxdep, rpvdep, 
rpmxdep, rpgt50, rpgt75, rpgt100, rpgt05, rpgt05x, rpgt10, rpgt10x, rpgt20,  
rpgt20x, rpxarea, rpvarea, rpmxar, areasum, rp100

Descriptions for all metrics are included in 
\emph{NRSA_Physical_Habitat_Metric_Descriptions.pdf} in the package
documentation.
}
\description{
This function calculates the residual pools portion 
of the physical habitat metrics for National Rivers and Streams 
Assessment (NRSA) data.  The function requires data frames containing 
the thalweg, channel geometry, and stream verification form data files.
}
\author{
Curt Seeliger \email{Seeliger.Curt@epa.gov}\cr
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
