% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRSAphab_indicators.r
\name{nrsaRiparianVegIndicator}
\alias{nrsaRiparianVegIndicator}
\title{Assign NRSA Riparian Vegetation indicator condition class}
\usage{
nrsaRiparianVegIndicator(x, sampID = "UID", ecoreg, protocol, xcmgw, lat, lon,
  area, elev, slope, xwidth)
}
\arguments{
\item{x}{A data frame containing, at minimum, the variables 
specified in the arguments for \emph{sampID}, ecoreg, realm, lrbs, and 
any variables necessary to calculate indicator condition classes 
for ecoregions in x. The necessary variables by realm and ecoreg 
are (none indicates no additional variables are necessary):
\itemize{
\item CPL, BOATABLE: lon, area
\item CPL, WADEABLE: lon
\item NAP, BOATABLE: lat
\item NAP, WADEABLE: area, xwidth
\item NPL, BOATABLE: lat, lon
\item NPL, WADEABLE: lat, lon, slope
\item SAP, BOATABLE: none
\item SAP, WADEABLE: area, elev
\item SPL, BOATABLE: lat, lon
\item SPL, WADEABLE: lon, elev
\item TPL, BOATABLE: none
\item TPL, WADEABLE: lon, elev
\item UMW, BOATABLE: lat, area, slope, xwidth
\item UMW, WADEABLE: slope, xwidth
\item WMT, BOATABLE: none
\item WMT, WADEABLE: area, elev, slope
\item XER, BOATABLE: none
\item XER, WADEABLE: area, slope, width
}}

\item{sampID}{A character vector containing the names of all 
variables in \emph{x} that specify a unique sample. If not specified, 
the default is \emph{UID}.}

\item{ecoreg}{A string with the name of the aggregated ecoregion variable. 
Valid values that correspond to regions used in NRSA are
CPL, NAP, NPL, SAP, SPL, TPL, UMW, WMTNS, and XER.}

\item{protocol}{A string with the name of the variable indicating
physical habitat protocol used. The expected values of this variable are 
'BOATABLE' and 'WADEABLE'.}

\item{xcmgw}{A string with the name of the variable for proportion mean(canopy + 
mid + ground riparian cover). Typically, the version used from NRSA 
physical habitat metrics is XCMGW.}

\item{lat}{A string with the name of the variable for latitude, assumed to 
be in decimal degrees using NAD83.}

\item{lon}{A string with the name of the variable for longitude, assumed to
be in decimal degrees using NAD83.}

\item{area}{A string with the name of the variable for watershed area in km2}

\item{elev}{A string with the name of the variable for elevation in meters. 
In NRSA, the variable used is ELEV_PT.}

\item{slope}{A string with the name of the variable for % stream slope. For 
NRSA, the variable used is XSLOPE.}

\item{xwidth}{A string with the name of the variable for mean wetted width
in meters. For NRSA, the variable used is XWIDTH.}
}
\value{
A data frame containing the variables in \emph{sampID} and RIPVEG_COND, 
the condition class for the Riparian Vegetation Cover indicator.
}
\description{
Using models based on NRSA 2008-9 and 2013-14 reference sites, 
this function determines the condition class by aggregated ecoregion and
sampling protocol used. The model used to determine the expected 
Log10(XCMGW + 0.01) (Mean canopy + mid + ground riparian cover) depends on 
the ecoregion and protocol.
}
\references{
Add Phil Kaufmann's technical report here
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
\keyword{survey}
