% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRSAphab_indicators.r
\name{nrsaRipDistIndicator}
\alias{nrsaRipDistIndicator}
\title{Assign NRSA Riparian Disturbance indicator
condition class}
\usage{
nrsaRipDistIndicator(x, sampID = "UID", w1_hall)
}
\arguments{
\item{x}{A data frame containing, at minimum, the variables 
specified in the arguments for sampID and w1_hall}

\item{sampID}{A character vector containing the names of all 
variables in inMets that specify a unique sample. If not specified, 
the default is \emph{UID}}

\item{w1_hall}{A string with the name of the variable representing all 
human disturbance. In NRSA, the name of the metric is W1_HALL.}
}
\value{
A data frame containing the variables in \emph{sampID} and RIPDIST_COND, 
the condition class for riparian disturbance.
}
\description{
Determine condition class based thresholds using 
the W1_HALL metric.
}
\references{
Add in from Phil Kaufmann later
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
\keyword{survey}
