% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrsaSlope}
\alias{nrsaSlope}
\title{Example data for NRSA channel geometry, e.g., slope and bearing measurements}
\format{
a data frame with 497 rows and 6 variables
\describe{
 \item{SITE}{Site visit identifier}
 \item{TRANSECT}{Transect value, range A-K}
 \item{SAMPLE_TYPE}{Sample type, corresponding to field form used, values
 PHAB_SLOPE and PHAB_CHANBFRONT}
 \item{REP}{Replicate for measurement}
 \item{PARAMETER}{Parameter name for field measurment}
 \item{VALUE}{Value of parameter}
}
}
\source{
aquamet package
}
\usage{
nrsaSlope
}
\description{
Example data for NRSA channel geometry, e.g., slope and bearing measurements
}
\keyword{datasets}
