% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrsaSubstrateEmbed.r
\name{nrsaSubstrateEmbed}
\alias{nrsaSubstrateEmbed}
\title{Calculate NRSA Substrate Embeddedness Metrics}
\usage{
nrsaSubstrateEmbed(percentEmbedded = NULL)
}
\arguments{
\item{percentEmbedded}{A data frame containing size class data 
for the dominant substrate class in the channel bottom, with the 
following columns:
\itemize{
     \item SITE integer or character specifying the site 
                visit
     \item TRANSECT character value specifying the transect
                    for which the value was recorded.
     \item ONBANK logical value specifying whether the values
                  were obtained at the bank or somewhere in the
                  channel
     \item VALUE character values
}}
}
\value{
Either a data frame when metric calculation is successful 
or a character string containing an error message when metric 
calculation is not successful.  The data frame contains the following 
columns:
\itemize{
   \item SITE - universal ID value
   \item METRIC - metric name
   \item VALUE - metric value
} 
Metrics calculated include: 
n33, n55, vcembed, vembed, xcembed, xembed
    
Descriptions for all metrics are included in 
\emph{NRSA_Physical_Habitat_Metric_Descriptions.pdf} in the package
documentation.
}
\description{
This function calculates the substrate 
embeddedness portion of the physical habitat metrics for 
National Rivers and Streams Assessment (NRSA) data.  The
function requires a data frame containing the channel crosssection 
data file.
}
\author{
Curt Seeliger \email{Seeliger.Curt@epa.gov}\cr
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
