% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrsaThalweg}
\alias{nrsaThalweg}
\title{Example data for NRSA thalweg-related measurements}
\format{
a data frame with 3065 rows and 6 variables
\describe{
 \item{SITE}{Site visit identifier}
 \item{TRANSECT}{Transect value, range A-K}
 \item{STATION}{Station between transects}
 \item{SAMPLE_TYPE}{Sample type, corresponding to field form used, values
 PHAB_THAL and PHAB_THALW}
 \item{PARAMETER}{Parameter name for field measurment}
 \item{VALUE}{Value of parameter}
}
}
\source{
ALL_THE_NRSA database
}
\usage{
nrsaThalweg
}
\description{
Example data for NRSA thalweg-related measurements
}
\keyword{datasets}
