% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrsaVisits}
\alias{nrsaVisits}
\title{Example data for NRSA visit information for example}
\format{
a data frame with 4 rows and 2 variables
\describe{
 \item{SITE}{Site visit identifier}
 \item{VALXSITE}{Sampling methodology used}
}
}
\source{
aquamet package
}
\usage{
nrsaVisits
}
\description{
Example data for NRSA visit information for example
}
\keyword{datasets}
