\name{visitsEx}
\alias{visitsEx}
\docType{data}
\title{Example Visits Dataset}
\description{
  A dataset containing visit data for use in physical habitat function examples.
}
\usage{data(visitsEx)}
\format{
  A data frame with 4 observations on the following 2 variables.
  \describe{
    \item{SITE}{unique site visit ID.}
    \item{VALXSITE}{X-site sampling status sub-category.}
  }
}
\details{
  These data are a small subset of the NRSA 2008-2009 visits dataset for example purposes only.
}
\examples{
data(visitsEx)
head(visitsEx)
}
\keyword{datasets}
