% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquamet-internal.r
\name{aquamet-internal}
\alias{aquamet-internal}
\alias{assignTaxCat}
\alias{calcSynCovers}
\alias{calcSynInfluence}
\alias{convert_to_char}
\alias{count}
\alias{dfCompare}
\alias{dfLengthen}
\alias{dfWiden}
\alias{expand.data.frame}
\alias{fillinDrawdownData}
\alias{fillinDrawdownData.expansion}
\alias{first}
\alias{gmean}
\alias{idr}
\alias{iqr}
\alias{interpolatePercentile}
\alias{is.subset}
\alias{intermediateMessage}
\alias{lag}
\alias{last}
\alias{lead}
\alias{modalClass}
\alias{modalClasses}
\alias{modalCount}
\alias{modalvalue}
\alias{modalValues}
\alias{normalizedCover}
\alias{nWadeableStationsPerTransect}
\alias{protectedMean}
\alias{protectedSum}
\alias{rename}
\alias{summaryby}
\alias{trimws}
\alias{uidCreate}
\alias{uidSeparate}
\alias{Dominance}
\alias{ShanDiversity}
\alias{tolindex}
\alias{tolindexFish}
\alias{aquametStandardizeArgument}
\alias{aquametStandardizeArgument.checkLegal}
\alias{aquametStandardizeArgument.checkRange}
\alias{aquametStandardizeArgument.checkStructure}
\alias{nlaStationInformation.islandStations}
\alias{nlaStationInformation.stationDepths}
\title{aquamet internal functions}
\usage{
aquametStandardizeArgument(arg, ..., ifdf=NULL, struct=list(SITE='integer', VALUE='double'), rangeLimits=NULL, legalValues=NULL, stopOnError=TRUE) 
aquametStandardizeArgument.checkLegal(arg, expectedLegal)
aquametStandardizeArgument.checkRange(arg, expectedRange)
aquametStandardizeArgument.checkStructure(arg, expectedStruct)
assignTaxCat(lt)
calcSynCovers(coverData, maxDrawdown, assumptions=FALSE)
calcSynInfluence(influenceData)
convert_to_char(df)
count(x)
dfCompare(df1, df2, byVars, zeroFudge=1e-17, verbose=FALSE)
dfLengthen(df, keys, name, value, values)
dfWiden(df, keys, name, values, makeNumeric=TRUE)
expand.data.frame(df, cols)
fillinDrawdownData(df, fillinValue='0', fillinHORIZ_DIST_DD='0')
fillinDrawdownData.expansion(df)
first(df, v, first.v)
gmean(x)
idr(x, method=2)
iqr(x, method=2)
interpolatePercentile(df, classVar, percentile, pctlVar, classBounds)
is.subset(a, b)
intermediateMessage(text, loc='middle', intermediateMessages=TRUE)
lag(df, v, lag.v, offset=1)
last(df, v, last.v)
lead(df, v, lead.v, offset=1)
modalClass(x, values, classes)
modalClasses(df, classes, weights, delim = ', ')
modalCount(x)
modalvalue(x, na.rm=FALSE)
modalValues(x, delim='&', na.rm=FALSE)
normalizedCover(df, coverValue, coverNorm, allowTotalBelow100=FALSE)
nWadeableStationsPerTransect(thal)
protectedMean(x, na.rm=FALSE, inf.rm=FALSE, nan.rm=FALSE, ...)
protectedSum(x, na.rm=FALSE, inf.rm=FALSE, nan.rm=FALSE, ...)
rename(df, old, new) 
summaryby(xxx,yyy,zzz)
trimws(text)
uidCreate(df, keys, sep='+')
uidSeparate(df, uidName, keyNames, sep='+')
Dominance(df, topN=1)
ShanDiversity(indata)
tolindex(indata, taxalist)
tolindexFish(indata, taxalist)
nlaStationInformation.islandStations(df)
nlaStationInformation.stationDepths(df)
}
\description{
These are internal functions to aquamet and not intended for use on their own.
}
\keyword{internal}
