% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignDistinct.r
\name{assignDistinct}
\alias{assignDistinct}
\title{Assign distinctness to each taxon}
\usage{
assignDistinct(cc, sampleID = "UID", taxlevels, final.name = NULL,
  special.taxa = NULL)
}
\arguments{
\item{cc}{Data frame containing all taxonomic fields specified
in the argument \emph{taxlevels} and the variables specified in 
the argument \emph{sampleID}.}

\item{sampleID}{A character vector containing the names of all 
variables in \emph{cc} that specify a unique sample. If not specified, 
the default is \emph{UID}}

\item{taxlevels}{A character vector containing the list of taxonomic
levels in data, in order from broadest to most specific}

\item{final.name}{An optional string containing the name of the column containing
the final taxon name for each record. This argument is only required if
the \emph{special.taxa} argument is provided.}

\item{special.taxa}{An optional character vector of taxon names that are considered
the lowest or target taxonomic level, even if they do not fit into the
hierarchical taxonomic level provided. These taxa are assigned an 
IS_DISTINCT value of 1.}
}
\value{
A data frame equivalent to the input data frame (cc) 
with IS_DISTINCT field added.
}
\description{
This function evaluates taxonomic distinctness 
within samples and assigns an indicator of distinctness 
(IS_DISTINCT) to each taxon (0 or 1) in each sample, based 
on distinctness within that sample.  Input data is expected 
to be count data with all counts above 0.
}
\examples{
\dontrun{
  data(distEx)
  head(distEx)
  # Assign distinctness to each taxon within samples
  distEx.1 <- assignDistinct(distEx,sampleID=c('UID','SAMPLE_TYPE'),
  taxlevels=c('PHYLUM','CLASS','ORDER','FAMILY','GENUS'),
  final.name='TARGET_TAXON',
  special.taxa=c('THIENEMANNIMYIA GENUS GR.', 'CERATOPOGONINAE', 'CRICOTOPUS/ORTHOCLADIUS'))
  
  head(distEx.1)
  }
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
\keyword{survey}
