% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignDistinct.r
\name{assignDistinct}
\alias{assignDistinct}
\title{Assign distinctness to each taxon}
\usage{
assignDistinct(cc, sampleID = "UID")
}
\arguments{
\item{cc}{Data frame containingtaxonomic fields for PHYLUM, 
CLASS, ORDER, FAMILY, and GENUS in addition to TAXA_ID, 
TARGET_TAXON, and the variables specified in the argument 
\emph{sampleID}.}

\item{sampleID}{A character vector containing the names of all 
variables in indf that specify a unique sample. If not specified, 
the default is \emph{UID}}
}
\value{
A data frame equivalent to the input data frame (cc) 
with IS_DISTINCT field added.
}
\description{
This function evaluates taxonomic distinctness 
within samples and assigns an indicator of distinctness 
(IS_DISTINCT) to each taxon (0 or 1) in each sample, based 
on distinctness within that sample.  Input data is expected 
to be count data with all counts above 0. The input data frame 
also needs to include TARGET_TAXON as the field containing the 
taxon name, TAXA_ID as the taxonomic identification number, and 
the following taxonomic fields fully populated (as much as 
possible, depending on taxonomic level of TARGET_TAXON: 
PHYLUM, CLASS, ORDER, FAMILY, and GENUS.

For the purposes of macroinvertebrates, the code assumes that 
taxa with TARGET_TAXON values of 'THIENEMANNIMYIA GENUS GR.', 
'CERATOPOGONINAE', or 'CRICOTOPUS/ORTHOCLADIUS' are at the lowest 
level and are assigned a value of 1 for IS_DISTINCT.
}
\examples{
\dontrun{
  data(distEx)
  head(distEx)
  # Assign distinctness to each taxon within samples
  distEx.1 <- assignDistinct(distEx)
  head(distEx.1)
  }
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
\keyword{survey}
