% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFishTolerMetricsSubset.r
\name{calcFishTolMets}
\alias{calcFishTolMets}
\title{Calculate all tolerance-related metrics}
\usage{
calcFishTolMets(indata, inTaxa = NULL, sampID = "UID",
  dist = "IS_DISTINCT", ct = "TOTAL", taxa_id = "TAXA_ID",
  tol = "TOLERANCE", tolval = "TOL_VAL", vel = "VELOCITY",
  habitat = "HABITAT", trophic = "TROPHIC", migr = "MIGRATORY",
  nonnat = "NONNATIVE")
}
\arguments{
\item{indata}{Input data frame containing variables as identified
in the arguments for \emph{sampID}, \emph{dist}, \emph{ct}, 
\emph{taxa_id}. If a variable for anomaly counts, as specified in 
the argument \emph{anom}, is included, additional metrics are 
calculated. The variable for the optional argument \emph{nonnat}
should also be included in this data frame.}

\item{inTaxa}{Data frame containing fish taxalist, along with autecology
traits. At a minimum, this taxalist must contain variables matching 
the argument for \emph{tol}. If additional traits are included, as 
specified by the arguments \emph{habitat}, \emph{trophic}, \emph{migr},
additional metrics that combine tolerance and other traits are also 
calculated. The variable specified in the argument \emph{taxa_id} 
must match that in \emph{indata}.}

\item{sampID}{sampID A character vector containing the names of all 
variables in indf that specify a unique sample. If not specified, 
the default is \emph{UID}.}

\item{dist}{A string with the name of the distinctness variable, 
which is assumed to have only values of 0 or 1. If not specified, 
the default is \emph{IS_DISTINCT}.}

\item{ct}{A string with the name of the count variable. If not 
specified, the default is \emph{TOTAL}.}

\item{taxa_id}{A string with the name of the taxon ID variable 
in \emph{indata} that matches that in \emph{inTaxa}. The default 
value is \emph{TAXA_ID}.}

\item{tol}{A string with the name of the variable in the 
\emph{inTaxa} taxalist containing tolerance categories. Valid
values include S (sensitive), I (intermediate), and T (tolerant).
The default value is \emph{TOLERANCE}.}

\item{tolval}{A string with the name of the variable in the 
\emph{inTaxa} taxalist containing numeric tolerance values.
Valid values range from 0 to 10. The default value is TOL_VAL.}

\item{vel}{A strings with the name of the variable in the 
\emph{inTaxa} taxalist containing velocity preference values.
Valid values include R (Rheophil), P (Pool), O (Other), or 
blank, if unknown. The default value is \emph{VELOCITY}.}

\item{habitat}{A string with the name of the variable in 
\emph{inTaxa} containing habitat preference values. Valid
values include B (benthic), W (water column), E (edge), or
blank, if unknown. The default value is \emph{HABITAT}.}

\item{trophic}{A string with the name of the variable in
\emph{inTaxa} containing trophic category values. Valid
values include I (invertivore), C (carnivore), O (omnivore),
H (herbivore), or blank if unknown. The default value is
\emph{inTaxa} is TROPHIC.}

\item{migr}{A string with the name of the variable in 
\emph{inTaxa} containing migratory status value. Valid 
values include N (No), Y (Yes), and blank if unknown. The
default value is \emph{MIGRATORY}.}

\item{nonnat}{A string with the name of the optional variable in 
\emph{inCts} containing non-native status. Valid values are 'Y' for 
non-native and 'N' for native. The default name 
is \emph{NONNATIVE}.}
}
\value{
A data frame containing the variables in sampID and 
the fish tolerance metrics as additional variables. Metric 
descriptions are included in \emph{NRSA_Fish_Metric_Descriptions.pdf},
included in this package. The names of
metrics include INTLNIND, INTLNTAX, INTLPIND, INTLPTAX, MTOLNIND, 
MTOLNTAX, MTOLPIND, MTOLPTAX, NTOLNIND, NTOLNTAX, NTOLPIND, 
NTOLPTAX, TOLRNIND, TOLRNTAX, TOLRPIND, TOLRPTAX, WTD_TV, TOTLNIND,
and TOTLNTAX.

If a non-native status variable is included, these metrics are also
calculated:
NAT_INTLNIND, NAT_INTLNTAX, NAT_INTLPIND, NAT_INTLPTAX, NAT_MTOLNIND, 
NAT_MTOLNTAX, NAT_MTOLPIND, NAT_MTOLPTAX, NAT_NTOLNIND, NAT_NTOLNTAX, 
NAT_NTOLPIND, NAT_NTOLPTAX,  NAT_TOLRNIND, NAT_TOLRNTAX, NAT_TOLRPIND, 
NAT_TOLRPTAX, NAT_WTD_TV, NAT_TOTLNTAX, NAT_TOTLNIND,
NAT_PIND, NAT_PTAX. 
 
Additional metrics calculated if the appropriate additional traits
are included in \emph{inTaxa}: INTLINVNIND, INTLINVNTAX, INTLINVPIND, 
INTLINVPTAX, INTLLOTNIND, INTLLOTNTAX, INTLLOTPIND, INTLLOTPTAX,
INTLMIGRNIND, INTLMIGRNTAX, INTLMIGRPIND, INTLMIGRPTAX,  INTLRHEONIND, 
INTLRHEONTAX, INTLRHEOPIND, INTLRHEOPTAX, NTOLBENTNIND, NTOLBENTNTAX, 
NTOLBENTPIND, NTOLBENTPTAX, NTOLCARNNIND, NTOLCARNNTAX, 
NTOLCARNPIND, NTOLCARNPTAX, NTOLINVNIND, NTOLINVNTAX, NTOLINVPIND, 
NTOLINVPTAX. 

If a non-native status variable is included, these additional metrics
are also calculated: NAT_INTLINVNIND, NAT_INTLINVNTAX, NAT_INTLINVPIND, 
NAT_INTLINVPTAX, NAT_INTLLOTNIND, NAT_INTLLOTNTAX, NAT_INTLLOTPIND, 
NAT_INTLLOTPTAX, NAT_INTLMIGRNIND, NAT_INTLMIGRNTAX, NAT_INTLMIGRPIND, 
NAT_INTLMIGRPTAX, NAT_INTLRHEONIND, NAT_INTLRHEONTAX, NAT_INTLRHEOPIND, 
NAT_INTLRHEOPTAX, NAT_NTOLBENTNIND, NAT_NTOLBENTNTAX, NAT_NTOLBENTPIND, 
NAT_NTOLBENTPTAX, NAT_NTOLCARNNIND, NAT_NTOLCARNNTAX, NAT_NTOLCARNPIND, 
NAT_NTOLCARNPTAX, NAT_NTOLINVNIND, NAT_NTOLINVNTAX, NAT_NTOLINVPIND, 
NAT_NTOLINVPTAX.
}
\description{
This function calculates all of the tolerance metrics,
and if additional trait values are included in the taxalist (habitat,
trophic, migr).
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
