% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlaAquaticMacrophytes.r
\name{nlaAquaticMacrophytes}
\alias{nlaAquaticMacrophytes}
\title{Calculate NLA Aquatic Macrophyte Metrics}
\usage{
nlaAquaticMacrophytes(emergent = NULL, floating = NULL, submergent = NULL,
  totalCover = NULL)
}
\arguments{
\item{emergent}{A data frame containing emergent vegetation class values for sites 
sampled with the columns:
\itemize{
\item SITE an integer or character value identifying a single site 
visit.
\item STATION a character value identifying the station within the SITE
\item VALUE an integer value, or character value that is castable to an 
integer, from 0-4 containing the emergent macrophyte cover
category.
}}

\item{floating}{A data frame containing floating vegetation class value for 
sites sampled with the columns:
\itemize{
\item SITE an integer or character value identifying a single site visit.
\item STATION a character value identifying the station within the SITE
\item VALUE an integer value, or character value that is castable to an 
integer, from 0-4 containing the floating macrophyte cover
category.
}}

\item{submergent}{A data frame containing submergent class value for sites 
sampled, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site visit.
\item STATION a character value identifying the station within the SITE
\item VALUE an integer value, or character value that is castable to an 
integer, from 0-4 containing the subemergent macrophyte cover
category.
}}

\item{totalCover}{A data frame containing total cover class value for sites 
sampled, with the columns:
\itemize{
\item SITE an integer or character value identifying a single site visit.
\item STATION a character value identifying the station within the SITE
\item VALUE an integer value, or character value that is castable to an 
integer, from 0-4 containing the total macrophyte cover
category.
}}
}
\value{
Either a data frame when metric calculation is successful or a 
character string containing an error message when metric calculation 
is not successful. The data frame contains the following columns:
\itemize{ 
    \item SITE - unique site visit identifier
    \item METRIC - metric name
    \item VALUE - metric value
      }
The output metrics include:
AMFPEMERGENT, AMFPFLOATING, AMFPSUBMERGENT, AMFPALL, AMFCEMERGENT,   
AMFCFLOATING, AMFCSUBMERGENT, AMFCALL, AMVEMERGENT, AMVFLOATING,    
AMVSUBMERGENT, AMVALL, AMNEMERGENT, AMNFLOATING, AMNSUBMERGENT, AMNALL,
AMIQALL, AMIDALL, AMITOTAL 

Descriptions for all metrics are included in 
\emph{NLA_Physical_Habitat_Metric_Descriptions.pdf} in the package
documentation.
}
\description{
This function calculates the aquatic macrophyte portion of the physical
habitat metrics for National Lakes Assessment (NLA) data.
}
\examples{
  head(nlaPhabEx)
  
  emerg <- subset(nlaPhabEx,PARAMETER=='AM_EMERGENT',select=-PARAMETER)
  float <- subset(nlaPhabEx,PARAMETER=='AM_FLOATING',select=-PARAMETER)
  submerg <- subset(nlaPhabEx,PARAMETER=='AM_SUBMERGENT',select=-PARAMETER)
  totcvr <- subset(nlaPhabEx,PARAMETER=='AM_TOTALCOVER',select=-PARAMETER)
  
  exAquMacro <- nlaAquaticMacrophytes(emergent=emerg,
  floating=float,submergent=submerg,totalCover=totcvr)
  
  head(exAquMacro)
 
}
\author{
Curt Seeliger \email{Seeliger.Curt@epa.gov}\cr
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
