% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLAphab_indicators.r
\name{nlaDrawdownIndicator}
\alias{nlaDrawdownIndicator}
\title{Calculate NLA Drawdown Indicator}
\usage{
nlaDrawdownIndicator(x, sampID, bfxVertDD, bfxHorizDD, ecoreg, lake_origin)
}
\arguments{
\item{x}{Input data frame containing other variables as specified in
arguments}

\item{sampID}{Vector of variables necessary to identify a unique sample 
in \emph{x}.}

\item{bfxVertDD}{Vertical height to highwater mark in meters}

\item{bfxHorizDD}{Horizontal distance to highwater mark in meters}

\item{ecoreg}{Lake ecoregion, based on aggregated Omernick ecoregions,
with valid values of CPL, NAP, NPL, SAP, SPL, TPL, UMW, WMT, XER.}

\item{lake_origin}{Lake origin, with valid values of 'NATURAL' 
or 'MAN_MADE'}
}
\value{
A data frame containing:
\itemize{
\item{sampID}{The variables in the argument \emph{sampID}}

\item{DRAWDOWN_COND}{Riparian anthropogenic disturbance condition class
 (Small/Medium/Large/Not Assessed)} 
}
}
\description{
Using metric values as inputs, calculate 
indicator score for lake drawdown, based on both horizontal
and vertical drawdown. Each is treated separately, and the 
two condition classes are then assessed to assign the 
larger of the two condition classes.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
\keyword{survey}
