% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrsaGeneral.r
\name{nrsaGeneral}
\alias{nrsaGeneral}
\title{Calculate NRSA General Metrics}
\usage{
nrsaGeneral(sampledTransects = NULL, sideChannels = NULL,
  transectSpacing = NULL, sideChannelTransects = c("XA", "XB", "XC", "XD",
  "XE", "XF", "XG", "XH", "XI", "XJ", "XK"))
}
\arguments{
\item{sampledTransects}{dataframe containing only the list of 
transects sampled for each site, one transect per row. Expected to 
contain the following columns:
 \itemize{
       \item SITE integer or character specifying the site 
                  visit
       \item TRANSECT character value specifying the transect
                      for which values were recorded.
}}

\item{sideChannels}{dataframe containing side channel presence data at each 
station, with the following columns:
\itemize{
         \item SITE integer or character specifying the site visit
         \item TRANSECT character value specifying the transect for which 
             values were recorded
         \item VALUE character values, expected to be 'Y', 'N' or NA
}}

\item{transectSpacing}{dataframe containing distances (in meters) between 
transects with the following columns:
\itemize{
     \item SITE integer or character specifying the site visit
     \item TRANSECT character value specifying the transect
                    for which the value was recorded.
     \item VALUE numeric values
}}

\item{sideChannelTransects}{character vector containing values of 
TRANSECT that are used to specify whether a transect is located on a 
side channel.  Default value is for EPA NARS use (XA, XB, etc.)
which occur parallel to main channel transects A, B, etc.}
}
\value{
Either a data frame when metric calculation is successful or a 
character string containing an error message when metric calculation is 
not successful.  The data frame contains the following columns:
\itemize{
\item SITE - universal ID value
\item METRIC - metric name
\item VALUE - metric value
}

Metrics calculated include: pct_side, reachlen, xtranspc, sidecnt

Descriptions for all metrics are included in 
\emph{NRSA_Physical_Habitat_Metric_Descriptions.pdf} in the package
documentation.
}
\description{
This function calculates the general portion of 
the physical habitat metrics for National Rivers and Streams 
Assessment (NRSA) data.  The function requires data frames containing 
the thalweg and channel geometry data files.
}
\examples{
head(thalwegEx)
head(changeomEx)

sampTr <- unique(thalwegEx[,c('SITE','TRANSECT')])
sideCh <- subset(thalwegEx,PARAMETER \%in\% c('SIDCHN','OFF_CHAN'))

# Creating the transect spacing for wadeable streams is more complicated 
# than for boatable streams because measurements are only made up to one 
# station before the end of the reach. 
wTr.1 <- subset(thalwegEx, PARAMETER=='INCREMNT' & TRANSECT=='A' & 
       STATION==0 & SAMPLE_TYPE=='PHAB_THALW',select=c('SITE','VALUE')) 

wTr.2 <- plyr::mutate(wTr.1, VALUE=as.numeric(VALUE)) 

wTr.3 <- merge(wTr.2, plyr::ddply(unique(subset(thalwegEx, SAMPLE_TYPE=='PHAB_THALW', 
                             select=c('SITE','TRANSECT','STATION'))),
             c('SITE','TRANSECT'), summarise, nSta = length(STATION)),
                                     by = 'SITE') 
                                     
wTr.4 <- plyr::ddply(wTr.3, c('SITE'), mutate, lastTran=max(TRANSECT)) 

wTr.5 <- plyr::mutate(wTr.4, VALUE = ifelse(TRANSECT!=lastTran, nSta*VALUE, (nSta-1)*VALUE)) 

wTr.6 <- dplyr::select(wTr.5, -nSta,-lastTran)

bTr <- subset(changeomEx,PARAMETER=='ACTRANSP',select=c('SITE','TRANSECT','VALUE'))

trDist <- rbind(wTr.6,bTr) \%>\% plyr::mutate(VALUE=as.numeric(VALUE))

generalOut <- nrsaGeneral(sampledTransects=sampTr, sideChannels=sideCh,
transectSpacing=trDist)

head(generalOut)
}
\author{
Curt Seeliger \email{Seeliger.Curt@epa.gov}\cr
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
